### 2026-02-20 #10088 @danielskatz
👋 @cvanaret - Thanks for your submission, which appears to be in scope for JOSS review. However, I note that you have some DOIs in your .bib file that need to be fixed to remove the prefix. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

Also, there's a section in the paper (State of the field) that's missing. Please add it, then use the command `@editorialbot generate pdf` to make a new PDF.

An editorialbot command needs to be the first entry in a new comment.
