+++ joss/paper.bib
@@ -7,7 +7,7 @@
   pages={25--57},
   year={2006},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s10107-004-0559-y}
+  doi={10.1007/s10107-004-0559-y}
 }
 
 @article{fletcher1998user,
@@ -25,7 +25,7 @@
   pages     = {35--59},
   year      = {2006},
   publisher = {Springer},
-  doi = {https://doi.org/10.1007/0-387-30065-1_4}
+  doi = {10.1007/0-387-30065-1_4}
 }
 
 @article{gill2005,
@@ -37,7 +37,7 @@
   pages     = {99--131},
   year      = {2005},
   publisher = {SIAM},
-  doi = {https://doi.org/10.1137/S0036144504446096}
+  doi = {10.1137/S0036144504446096}
 }
 
 @unpublished{VanaretLeyffer2024,
@@ -52,7 +52,7 @@
   author    = {Nocedal, Jorge and Wright, Stephen J},
   year      = {2006},
   publisher = {Springer},
-  doi = {https://doi.org/10.1007/978-0-387-40065-5}
+  doi = {10.1007/978-0-387-40065-5}
 }
 
 @book{lee2011,
@@ -60,7 +60,7 @@
   author    = {Lee, Jon and Leyffer, Sven},
   year      = {2011},
   publisher = {Springer Science \& Business Media},
-  doi = {https://doi.org/10.1007/978-1-4614-1927-3}
+  doi = {10.1007/978-1-4614-1927-3}
 }
 
 @article{fourer1990,
@@ -71,7 +71,7 @@
   number  = {5},
   pages   = {519--554},
   year    = {1990},
-  doi = {https://doi.org/10.1007/978-3-642-83724-1_12}
+  doi = {10.1007/978-3-642-83724-1_12}
 }
 
 @techreport{gay1997hooking,
@@ -97,5 +97,5 @@
   pages={123--160},
   year={1995},
   publisher={ACM New York, NY, USA},
-  doi = {https://doi.org/10.1145/200979.201043}
+  doi = {10.1145/200979.201043}
 }
