### 2026-02-20 d740966
Moved some of the text to State of the field + added two references to MECO cooperation
