### 2026-02-20 38f5630
Moved some of the text to State of the field + added two references to MECO cooperation
