+++ joss/paper.bib
@@ -99,3 +99,23 @@
   publisher={ACM New York, NY, USA},
   doi = {10.1145/200979.201043}
 }
+
+@inproceedings{kiessling2024almost,
+  title={An almost feasible sequential linear programming algorithm},
+  author={Kiessling, David and Vanaret, Charlie and Astudillo, Alejandro and Decr{\'e}, Wilm and Swevers, Jan},
+  booktitle={2024 European Control Conference (ECC)},
+  pages={2328--2335},
+  year={2024},
+  organization={IEEE},
+  doi = {10.23919/ECC64448.2024.10590864}
+}
+
+@article{kiessling2025unified,
+  title={A unified funnel restoration SQP algorithm: D. Kiessling et al.},
+  author={Kiessling, David and Leyffer, Sven and Vanaret, Charlie},
+  journal={Mathematical Programming},
+  pages={1--45},
+  year={2025},
+  publisher={Springer},
+  doi = {10.1007/s10107-025-02284-3}
+}
