### 2026-02-21 413371e
Moved some stuff around and reformulated some other stuff
