+++ joss/paper.md
@@ -37,22 +37,22 @@ As of February 2026, Uno supports sequential (convex and nonconvex) quadratic pr
 
 Uno breaks down optimization algorithms into reusable modular components such as step computation, constraint reformulation, globalization techniques, and acceptance criteria.
 This allows classical and hybrid methods to be configured and compared within a single framework.
+For full mathematical details of the algorithms implemented in Uno, see [@VanaretLeyffer2024].
 
 The core C\texttt{++} code of Uno is organized into modular, object-oriented components that separate the mathematical logic of the algorithms from implementation details such as memory management, data structures, and computational routines.
-For full mathematical details of the algorithms implemented in Uno, see [@VanaretLeyffer2024].
 Uno provides interfaces to Julia, Python, C, Fortran, and AMPL, enabling use across scientific computing environments.
-Precompiled artifacts are available on GitHub, and the solver is directly available via `UnoSolver.jl` in Julia or `unopy` in Python.
+Precompiled artifacts are available on GitHub, and the solver can be accessed directly via `UnoSolver.jl` in Julia or `unopy` in Python.
 
 # Statement of need
 
 Nonlinearly constrained optimization is central to engineering, optimal control, machine learning, and scientific modeling [@nocedal2006].
 It also plays a key role in mixed-integer nonlinear optimization [@lee2011].
 
-Typical nonlinear solvers rely on recurring ingredients such as step computation, constraint handling, and globalization strategies.
-These ingredients appear across major paradigms including sequential quadratic programming, interior-point methods, and augmented Lagrangian methods, but they are usually developed as separate solver families.
-Algorithmic ideas are typically evaluated at the level of complete solvers, rather than individual components.
+Typical nonlinear solvers share common building blocks such as step computation, constraint handling, and globalization strategies.
+These components can be found across major paradigms such as sequential quadratic programming, interior-point methods, and augmented Lagrangian methods, but these are usually developed as separate solver families.
+Algorithmic ideas are typically tested at the level of complete solvers, rather than individual components.
 
-A framework that exposes algorithmic components explicitly within a consistent architecture is therefore needed.
+There is therefore a need for a framework that makes algorithmic components explicit within a unified and consistent architecture.
 
 # State of the field
 
@@ -60,31 +60,30 @@ Popular solvers such as IPOPT [@wachter2006implementation], KNITRO [@byrd2006], 
 These solvers act mainly as black-box engines.
 Testing new algorithmic variants usually requires intrusive modification or reimplementation.
 
-This motivates the development of a framework built from composable algorithmic building blocks.
+This motivates the development of a framework built upon composable, algorithmic building blocks.
 
 # Unification framework
 
-Uno addresses the lack of component-level experimentation by providing a composable optimization framework in which algorithms are constructed from modular components corresponding to mathematical concepts such as step computation, constraint reformulation, and globalization techniques.
+Uno enables experimentation at the component level by providing a composable optimization framework in which algorithms are constructed from modular components that reflect mathematical concepts such as step computation, constraint reformulation, and globalization techniques.
 
-Within this framework, strategies are organized into a coherent hierarchy, illustrated in the wheel of strategies (\autoref{fig:wheel}): the outer ring represents high-level layers, the middle ring represents algorithmic ingredients that are automatically combined, and the inner ring lists possible strategies for each ingredient.
+Within this framework, strategies are organized into a coherent hierarchy, illustrated in the wheel of strategies (\autoref{fig:wheel}): the outer ring represents high-level layers, the middle ring represents algorithmic ingredients that are combined automatically, and the inner ring lists possible strategies for each ingredient.
[…]
