### 2026-02-23 05dbac9
Updated bibtex entry [skip ci]
Updated bibtex entry with new year and note.
