### 2026-02-23 8a4d559
Simplified update of L, Ltilde and V, knowing that we modify only the current column of S and Y
