+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.cpp
@@ -71,7 +71,7 @@ namespace uno {
 
    void LBFGSHessian::notify_accepted_iterate(Statistics& statistics, const Iterate& current_iterate, const Iterate& trial_iterate,
          EvaluationCache& evaluation_cache) {
-      DEBUG << "\n*** Updating the BFGS memory at slot " << this->current_memory_slot << '\n';
+      DEBUG << "\n*** Updating the BFGS memory at slot " << this->current_index << '\n';
       // update the matrices S, Y and D
       this->update_S(current_iterate, trial_iterate);
       this->update_Y(current_iterate, trial_iterate, evaluation_cache);
@@ -82,8 +82,8 @@ namespace uno {
 
       // check that the latest D entry s^T y is > 0
       // TODO implement Procedure 18.2 (Damped BFGS Updating) from Numerical Optimization
-      if (0. < this->D[this->current_memory_slot]) {
-         DEBUG << "S, Y and D updated at slot " << this->current_memory_slot << '\n';
+      if (0. < this->D[this->current_index]) {
+         DEBUG << "S, Y and D updated at slot " << this->current_index << '\n';
          this->number_entries_in_memory = std::min(this->number_entries_in_memory + 1, this->memory_size);
          // notify_accepted_iterate is called at the end of a major iteration. At this point, we don't know yet whether
          // the L-BFGS Hessian will be used. We therefore delay the update to the beginning of the next major iteration
@@ -145,7 +145,7 @@ namespace uno {
 
    void LBFGSHessian::update_S(const Iterate& current_iterate, const Iterate& trial_iterate) {
       // TODO check that the S entry isn't too small
-      this->S.column(this->current_memory_slot) = view(trial_iterate.primals, 0, this->model.number_variables) -
+      this->S.column(this->current_index) = view(trial_iterate.primals, 0, this->model.number_variables) -
          view(current_iterate.primals, 0, this->model.number_variables);
    }
    
@@ -156,63 +156,53 @@ namespace uno {
          evaluation_cache.current_evaluations, this->current_lagrangian_gradient);
       this->model.evaluate_lagrangian_gradient(trial_iterate.primals, trial_iterate.multipliers, this->fixed_objective_multiplier,
          evaluation_cache.trial_evaluations, this->trial_lagrangian_gradient);
-      this->Y.column(this->current_memory_slot) = this->trial_lagrangian_gradient - this->current_lagrangian_gradient;
+      this->Y.column(this->current_index) = this->trial_lagrangian_gradient - this->current_lagrangian_gradient;
    }
 
    void LBFGSHessian::update_D() {
-      this->D[this->current_memory_slot] = dot(this->S.column(this->current_memory_slot),
-         this->Y.column(this->current_memory_slot));
+      this->D[this->current_index] = dot(this->S.column(this->current_index), this->Y.column(this->current_index));
    }
 
    void LBFGSHessian::recompute_hessian_representation() {
-      // TODO figure out if we're extending or replacing in memory. But I think we don't have to with out floating slot index
       assert(0 < this->number_entries_in_memory);
-      assert(0 < this->D[this->current_memory_slot]);
+      assert(0 < this->D[this->current_index]);
 
       DEBUG << "\n*** Recomputing the Hessian representation with " << this->number_entries_in_memory << " entries\n";
-      // DEBUG << "this->current_memory_slot = " << this->current_memory_slot << '\n';
 
       // update invsqrt_D = 1/sqrt_D
-      this->invsqrt_D[this->current_memory_slot] = 1./std::sqrt(this->D[this->current_memory_slot]);
+      this->invsqrt_D[this->current_index] = 1./std::sqrt(this->D[this->current_index]);
 
[…]
