### 2026-02-23 #10088 @danielskatz
Thanks @cvanaret - if you can suggest a few more potential reviews, this will help the editor. Note that you can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
