### 2026-02-23 017fdde
Use BLAS for `Vector::scale` (#528)
* Use BLAS to scale Vector

* Removed unused functions in SparseVector
