### 2026-02-24 #10088 @amontoison
- Anugrah Jo Joshy (https://reviewers.joss.theoj.org/reviewers/4460)
- Justin Gray (https://reviewers.joss.theoj.org/reviewers/1613)

They may also be a good reviewers.
