### 2026-02-24 e6aeced
Introduced BLASVector and MutableBLASVector, inherited by Vector and VectorView
