+++ uno/linear_algebra/Vector.hpp
@@ -9,12 +9,12 @@
 #include <string>
 #include <vector>
 #include <initializer_list>
-#include "linear_algebra/BLAS.hpp"
+#include "BLASVector.hpp"
 #include "symbolic/Range.hpp"
 
 namespace uno {
    template <typename ElementType>
-   class Vector {
+   class Vector: public MutableBLASVector<ElementType> {
    public:
       using value_type = ElementType;
       // iterators
@@ -27,7 +27,7 @@ namespace uno {
       Vector(std::initializer_list<ElementType> initializer_list): vector(initializer_list) { }
       Vector(const Vector<ElementType>& other) noexcept : vector(other.vector) { }
       Vector(Vector<ElementType>&& other) noexcept : vector(std::move(other.vector)) { }
-      ~Vector() = default;
+      ~Vector() override = default;
 
       // copy assignment operator
       Vector<ElementType>& operator=(const Vector<ElementType>& other) {
@@ -56,17 +56,34 @@ namespace uno {
          return *this;
       }
 
-      // random access
-      ElementType& operator[](size_t index) { return this->vector[index]; }
-      const ElementType& operator[](size_t index) const { return this->vector[index]; }
+      [[nodiscard]] size_t size() const override {
+         return this->vector.size();
+      }
 
-      // size
-      [[nodiscard]] size_t size() const { return this->vector.size(); }
-      [[nodiscard]] bool empty() const { return (this->size() == 0); }
+      ElementType* data() override {
+         return this->vector.data();
+      }
+
+      const ElementType* data() const override {
+         return this->vector.data();
+      }
+
+      ElementType& operator[](size_t index) override {
+         return this->vector[index];
+      }
+
+      const ElementType& operator[](size_t index) const override {
+         return this->vector[index];
+      }
 
       // size and capacity
-      void reserve(size_t new_capacity) { this->vector.reserve(new_capacity); }
-      void resize(size_t new_size) { this->vector.resize(new_size); }
+      void reserve(size_t new_capacity) {
+         this->vector.reserve(new_capacity);
+      }
+
+      void resize(size_t new_size) {
+         this->vector.resize(new_size);
+      }
 
       // iterators
       iterator begin() noexcept { return this->vector.begin(); }
@@ -83,19 +100,6 @@ namespace uno {
             this->vector[index] = value;
          }
       }
-
-      void scale(ElementType factor) {
-         const int size = static_cast<int>(this->size());
-         constexpr int increment = 1;
-         BLAS_scale_vector(&size, &factor, this->vector.data(), &increment);
-      }
-
-      void operator*=(ElementType factor) {
-         this->scale(factor);
-      }
-
-      ElementType* data() { return this->vector.data(); }
-      const ElementType* data() const { return this->vector.data(); }
 
       void print(std::ostream& stream) const {
          for (const ElementType& element: *this) {
@@ -122,14 +126,6 @@ namespace uno {
       return stream;
    }
 
-   // subtract operator
-   template <typename ResultExpression, typename Expression>
-   void operator-=(ResultExpression&& result, const Expression& expression) {
-      for (size_t index: Range(result.size())) {
-         result[index] -= expression[index];
-      }
-   }
-
    template <typename Container>
    std::string join(const Container& vector, const std::string& separator) {
       std::string result{};
@@ -144,10 +140,14 @@ namespace uno {
       return result;
    }
 
-   inline double dot(const Vector<double>& x, const Vector<double>& y) {
-      const int size = static_cast<int>(std::min(x.size(), y.size()));
-      constexpr int increment = 1;
-      return BLAS_dot_product(&size, x.data(), &increment, y.data(), &increment);
[…]
