+++ uno/symbolic/VectorView.hpp
@@ -1,60 +1,0 @@
-// Copyright (c) 2018-2024 Charlie Vanaret
-// Licensed under the MIT license. See LICENSE file in the project directory for details.
-
-#ifndef UNO_VECTORVIEW_H
-#define UNO_VECTORVIEW_H
-
-namespace uno {
-   // span of an arbitrary container: allocation-free view of a certain length
-   template <typename Vector>
-   class VectorView {
-   public:
-      using value_type = typename std::remove_reference_t<Vector>::value_type;
-
-      VectorView(Vector&& vector, size_t start, size_t end):
-            vector(std::forward<Vector>(vector)), start(start), end(std::min(end, vector.size())) {
-         if (end < start) {
-            throw std::runtime_error("The view ends before its starting point");
-         }
-      }
-
-      template <typename E>
-      VectorView& operator=(const E& expression) {
-         assert(this->size() == expression.size() && "The view and the expression have different sizes");
-
-         for (size_t index: Range(this->size())) {
-            this->vector[this->start + index] = expression[index];
-         }
-         return *this;
-      }
-
-      // preconditions: expression != nullptr, i < length
-      [[nodiscard]] value_type& operator[](size_t index) noexcept { return this->vector[this->start + index]; }
-      [[nodiscard]] value_type operator[](size_t index) const noexcept { return this->vector[this->start + index]; }
-      [[nodiscard]] size_t size() const noexcept { return this->end - this->start; }
-
-      // [[nodiscard]] iterator begin() const noexcept { return iterator(*this, 0); }
-      // [[nodiscard]] iterator end() const noexcept { return iterator(*this, this->length); }
-
-   protected:
-      Vector vector;
-      const size_t start;
-      const size_t end;
-   };
-
-   // free function
-   template <typename Expression>
-   VectorView<Expression> view(Expression&& expression, size_t start, size_t end) {
-      return {std::forward<Expression>(expression), start, end};
-   }
-
-   template <typename Expression>
-   std::ostream& operator<<(std::ostream& stream, const VectorView<Expression>& view) {
-      for (size_t index: Range(view.size())) {
-         stream << view[index] << " ";
-      }
-      return stream;
-   }
-} // namespace
-
-#endif //UNO_VECTORVIEW_H
