### 2026-02-24 #10088 @drvinceknight
@victoraalves @anugrahjo @JustinSGray I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “Uno: a composable framework for nonlinearly constrained optimization”. You can find more information at the top of this Github issue.
