### 2026-02-24 4971a73
Restored the barrier method (a line in Subproblem::assemble_augmented_rhs was commented out)
