### 2026-02-24 d0e3927
MutableVectorView: load operator of superclass with "using"
