### 2026-02-24 abf1388
`Vector` and `VectorView` inherit from `BLASVector` (#529)
* Introduced BLASVector and MutableBLASVector, inherited by Vector and VectorView

* Restored the barrier method (a line in Subproblem::assemble_augmented_rhs was commented out)

* Removed useless dot()

* MutableVectorView: load operator of superclass with "using"
