+++ uno/ingredients/subproblem_solvers/BQPD/BQPDSolver.cpp
@@ -7,11 +7,11 @@
 #include "ingredients/subproblem/Subproblem.hpp"
 #include "linear_algebra/Indexing.hpp"
 #include "linear_algebra/Vector.hpp"
+#include "linear_algebra/VectorView.hpp"
 #include "optimization/Direction.hpp"
 #include "optimization/Iterate.hpp"
 #include "optimization/WarmstartInformation.hpp"
 #include "options/Options.hpp"
-#include "symbolic/VectorView.hpp"
 #include "tools/Logger.hpp"
 #include "fortran_interface.h"
 
