### 2026-02-24 17835b8
Adapted LBFGSHessian to support BLASVector and subclasses Vector and VectorView. Implemented special operation for y += a * x with BLAS
