### 2026-02-25 e0b345f
Added MutableBLASVector::operator=(Subtraction) and simplified syntax in LBFGSHessian and PrimalDualInteriorPointProblem
