+++ uno/linear_algebra/BLASVector.hpp
@@ -9,6 +9,7 @@
 #include "BLAS.hpp"
 #include "symbolic/Range.hpp"
 #include "symbolic/ScalarMultiple.hpp"
+#include "symbolic/Subtraction.hpp"
 
 namespace uno {
    // constant contiguous array in memory on which BLAS can be called
@@ -34,7 +35,7 @@ namespace uno {
    public:
       MutableBLASVector() = default;
 
-      // specialized operator= when the other vector has the member function data()
+      // specialized operation y = x, when the other vector has the member function data()
       template <typename Vector, decltype(Vector{}.data()) = true>
       auto operator=(const Vector& other) {
          assert(other.size() <= this->size() && "The other vector is larger than the current vector");
@@ -44,17 +45,38 @@ namespace uno {
          return *this;
       }
 
-      // generic operator= for arbitrary expressions
-      template <typename Expression>
-      MutableBLASVector& operator=(const Expression& expression) {
-         static_assert(std::is_same_v<typename Expression::value_type, T>);
-         assert(expression.size() <= this->size() && "The expression is larger than the current vector");
-         for (size_t index: Range(expression.size())) {
-            this->operator[](index) = expression[index];
+      // generic operation y = x
+      template <typename Vector>
+      MutableBLASVector& operator=(const Vector& other) {
+         static_assert(std::is_same_v<typename Vector::value_type, T>);
+         assert(other.size() <= this->size() && "The expression is larger than the current vector");
+         for (size_t index: Range(other.size())) {
+            this->operator[](index) = other[index];
          }
          return *this;
       }
 
+      // specialized operation y = a * x
+      // note: no BLAS operation available
+      template <typename Vector>
+      MutableBLASVector& operator=(ScalarMultiple<Vector>&& other) {
+         assert(other.get_expression().size() <= this->size() && "The other vector is larger than the current vector");
+         for (size_t index: Range(this->size())) {
+            this->operator[](index) = other.get_factor() * other.get_expression()[index];
+         }
+         return *this;
+      }
+
+      // specialized operation y = x - z
+      // note: no BLAS operation available
+      template <typename Vector>
+      MutableBLASVector& operator=(Subtraction<Vector, Vector>&& other) {
+         *this = other.get_expression1();
+         *this -= other.get_expression2();
+         return *this;
+      }
+
+      // specialized operation y += a * x
       template <typename Vector>
       MutableBLASVector& operator+=(ScalarMultiple<Vector>&& other) {
          assert(other.size() <= this->size() && "The other vector is larger than the current vector");
@@ -65,6 +87,7 @@ namespace uno {
          return *this;
       }
 
+      // generic operation y += x
       template <typename Vector>
       MutableBLASVector& operator+=(const Vector& other) {
          assert(other.size() <= this->size() && "The other vector is larger than the current vector");
