### 2026-02-25 a7c5059
Removed useless data structures in COOWorkspace + modified signature of OptimizationProblem::evaluate_constraints
