### 2026-02-25 9fdc86e
Removed useless data structures in COOWorkspace + modified signature of OptimizationProblem::evaluate_constraints (#530)
