### 2026-02-25 6dcd948
Removed useless data structures in COOWorkspace + modified signature of OptimizationProblem::evaluate_constraints (#530)
