### 2026-02-25 75a9fef
Rewrote OptimizationProblem::evaluate_lagrangian_gradient with operator overloading, calls BLAS under the hood
