### 2026-02-25 68627ab
Added specialized (if .data() is available) and generic operator= in BLASVector
