### 2026-02-25 cc85c0c
`MutableBLASVector`: pick specialized `operator=` if `.data()` is available (#531)
* Rewrote OptimizationProblem::evaluate_lagrangian_gradient with operator overloading, calls BLAS under the hood

* Added specialized (if .data() is available) and generic operator= in BLASVector

* Removed useless include
