### 2026-02-26 b4fe7dd
Deleted unused DenseColumn
