### 2026-02-26 752476f
Use overloading of DenseMatrix::operator= to handle U = delta S + V * L_invsqrt_D^T + call to dgemm
