+++ uno/linear_algebra/DenseMatrix.hpp
@@ -7,31 +7,68 @@
 #include <cassert>
 #include <ostream>
 #include <vector>
+#include <iostream>
+#include "linear_algebra/BLAS.hpp"
 #include "VectorView.hpp"
+#include "symbolic/Multiplication.hpp"
 #include "symbolic/Range.hpp"
+#include "symbolic/Sum.hpp"
+#include "symbolic/Transpose.hpp"
 
 namespace uno {
-   enum class MatrixShape {
-      GENERAL,
-      LOWER_TRIANGULAR,
-      UPPER_TRIANGULAR
-   };
-
    // DenseMatrix is an m x n matrix in column-major order where the columns are concatenated in a long vector
-   template <typename ElementType, MatrixShape Shape = MatrixShape::GENERAL>
+   template <typename ElementType>
    class DenseMatrix {
    public:
       using value_type = ElementType;
 
       DenseMatrix(size_t number_rows, size_t number_columns);
-      DenseMatrix() = default;
-      DenseMatrix(const DenseMatrix& other);
-      DenseMatrix& operator=(const DenseMatrix& other) = default;
-      DenseMatrix& operator=(DenseMatrix&& other) noexcept = default;
+
+      // copy an existing matrix into this object
+      DenseMatrix& operator=(DenseMatrix& other) {
+         assert(other.number_rows == this->number_rows && other.number_columns == this->number_columns);
+         const int size = static_cast<int>(this->number_rows * this->number_columns);
+         constexpr int increment = 1;
+         BLAS_copy_vector(&size, other.data(), &increment, this->data(), &increment);
+         return *this;
+      }
+
       ~DenseMatrix() = default;
 
-      [[nodiscard]] size_t get_number_rows() const;
-      [[nodiscard]] size_t get_number_columns() const;
+      // specialized operator= for C := beta * C + A*B^T
+      // use different matrix types in case one of them has a different type (e.g., submatrix)
+      template <typename Matrix1, typename Matrix2, typename Matrix3>
+      DenseMatrix& operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2, Transpose<Matrix3>>>&& expression) {
+         const double beta = expression.get_left().get_factor();
+         auto& C = expression.get_left().get_expression();
+         auto& A = expression.get_right().get_left();
+         auto& B = expression.get_right().get_right().get_matrix();
+         // check that "this" is indeed on the left side of the sum
+         if (&C != this) {
+            throw std::runtime_error("DenseMatrix::operator= is not defined when C != this");
+         }
+         if (this->number_rows != A.number_rows || this->number_columns != B.number_rows) {
+            throw std::runtime_error("Dimension mismatch in DenseMatrix::operator=");
+         }
+         /*
+         std::cout << "C/this has size (" << this->number_rows << ", " << this->number_columns << ")\n";
+         std::cout << "A has size (" << A.number_rows << ", " << A.number_columns << ")\n";
+         std::cout << "B has size (" << B.number_rows << ", " << B.number_columns << ")\n";
+         */
+         const char transa = 'N'; // A
+         const char transb = 'T'; // B^T
+         const int m = static_cast<int>(A.number_rows); // number of rows of A
+         const int n = static_cast<int>(B.number_rows); // number of columns of B^T
+         const int k = static_cast<int>(A.number_columns); // number of columns of A
+         constexpr double alpha = 1.;
+         const int lda = static_cast<int>(A.number_rows); // leading dimension of A
+         const int ldb = static_cast<int>(B.number_rows); // leading dimension of B
+         const int ldc = static_cast<int>(this->number_rows); // leading dimension of C/this
[…]
