### 2026-02-26 c106d05
Added missing "const T* DenseMatrix::data() const"!
