### 2026-02-26 c586043
Use operator*= overloading to handle U = U J⁻ᵀ as U *= transpose(inverse(J)) 🤩
