+++ uno/linear_algebra/DenseMatrix.hpp
@@ -7,9 +7,9 @@
 #include <cassert>
 #include <ostream>
 #include <vector>
-#include <iostream>
 #include "linear_algebra/BLAS.hpp"
 #include "VectorView.hpp"
+#include "symbolic/Inverse.hpp"
 #include "symbolic/Multiplication.hpp"
 #include "symbolic/Range.hpp"
 #include "symbolic/Sum.hpp"
@@ -17,115 +17,140 @@
 
 namespace uno {
    // DenseMatrix is an m x n matrix in column-major order where the columns are concatenated in a long vector
-   template <typename ElementType>
+   template <typename T>
    class DenseMatrix {
    public:
-      using value_type = ElementType;
+      using value_type = T;
 
       DenseMatrix(size_t number_rows, size_t number_columns);
-      
-      // copy an existing matrix into this object
-      DenseMatrix& operator=(const DenseMatrix& other) {
-         assert(other.number_rows == this->number_rows && other.number_columns == this->number_columns);
-         const int size = static_cast<int>(this->number_rows * this->number_columns);
-         constexpr int increment = 1;
-         BLAS_copy_vector(&size, other.data(), &increment, this->data(), &increment);
-         return *this;
-      }
-
+      DenseMatrix& operator=(const DenseMatrix& other);
       ~DenseMatrix() = default;
 
       // specialized operator= for C := beta * C + A*B^T
-      // use different matrix types in case one of them has a different type (e.g., submatrix)
       template <typename Matrix1, typename Matrix2, typename Matrix3>
-      DenseMatrix& operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2, Transpose<Matrix3>>>&& expression) {
-         const double beta = expression.get_left().get_factor();
-         const auto& C = expression.get_left().get_expression();
-         const auto& A = expression.get_right().get_left();
-         const auto& B = expression.get_right().get_right().get_matrix();
-         // check that "this" is indeed on the left side of the sum
-         if (&C != this) {
-            throw std::runtime_error("DenseMatrix::operator= is not defined when C != this");
-         }
-         if (this->number_rows != A.number_rows || this->number_columns != B.number_rows) {
-            throw std::runtime_error("Dimension mismatch in DenseMatrix::operator=");
-         }
-         /*
-         std::cout << "C/this has size (" << this->number_rows << ", " << this->number_columns << ")\n";
-         std::cout << "A has size (" << A.number_rows << ", " << A.number_columns << ")\n";
-         std::cout << "B has size (" << B.number_rows << ", " << B.number_columns << ")\n";
-         */
-         const char transa = 'N'; // A
-         const char transb = 'T'; // B^T
-         const int m = static_cast<int>(A.number_rows); // number of rows of A
-         const int n = static_cast<int>(B.number_rows); // number of columns of B^T/number of rows of B
-         const int k = static_cast<int>(A.number_columns); // number of columns of A
-         constexpr double alpha = 1.;
-         const int lda = static_cast<int>(A.number_rows); // leading dimension of A
-         const int ldb = static_cast<int>(B.number_rows); // leading dimension of B
-         const int ldc = static_cast<int>(this->number_rows); // leading dimension of C/this
-         BLAS_matrix_matrix_product(&transa, &transb, &m, &n, &k, &alpha, A.data(), &lda, B.data(), &ldb, &beta,
-            this->data(), &ldc);
-         return *this;
-      }
+      DenseMatrix& operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2, Transpose<Matrix3>>>&& expression);
 
[…]
