### 2026-02-26 a8f1c67
Added BLASMatrix as superclass of DenseMatrix and Submatrix + wrote U = (δ S + Y D⁻¹ Lᵀ) J⁻ᵀ with submatrices
