+++ uno/linear_algebra/DenseMatrix.hpp
@@ -4,48 +4,66 @@
 #ifndef UNO_DENSEMATRIX_H
 #define UNO_DENSEMATRIX_H
 
-#include <cassert>
 #include <ostream>
 #include <vector>
-#include "linear_algebra/BLAS.hpp"
+#include "linear_algebra/BLASMatrix.hpp"
 #include "VectorView.hpp"
-#include "symbolic/Inverse.hpp"
-#include "symbolic/Multiplication.hpp"
 #include "symbolic/Range.hpp"
-#include "symbolic/Sum.hpp"
-#include "symbolic/Transpose.hpp"
 
 namespace uno {
    // DenseMatrix is an m x n matrix in column-major order where the columns are concatenated in a long vector
    template <typename T>
-   class DenseMatrix {
+   class DenseMatrix: public BLASMatrix<T> {
    public:
       using value_type = T;
 
+      class Submatrix: public BLASMatrix<T> {
+      public:
+         using value_type = T;
+
+         Submatrix(DenseMatrix<T>& matrix, size_t number_rows, size_t number_columns):
+               BLASMatrix<T>(number_rows, number_columns, matrix.number_rows),
+               matrix(matrix) {
+         }
+
+         // operators
+         using BLASMatrix<T>::operator=;
+         using BLASMatrix<T>::operator*=;
+         // delegate copy operator to that of BLASMatrix
+         Submatrix& operator=(const Submatrix& other) {
+            BLASMatrix<T>::operator=(other);
+            return *this;
+         }
+
+         [[nodiscard]] T* data() override {
+            return this->matrix.data();
+         }
+
+         [[nodiscard]] const T* data() const override {
+            return this->matrix.data();
+         }
+
+         DenseMatrix<T>& matrix;
+      };
+
       DenseMatrix(size_t number_rows, size_t number_columns);
-      DenseMatrix& operator=(const DenseMatrix& other);
-      ~DenseMatrix() = default;
+      ~DenseMatrix() override = default;
 
-      // specialized operator= for C := beta * C + A*B^T
-      template <typename Matrix1, typename Matrix2, typename Matrix3>
-      DenseMatrix& operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2, Transpose<Matrix3>>>&& expression);
-
-      // specialized operator= for B := B A⁻ᵀ
-      template <typename Matrix>
-      DenseMatrix& operator*=(Transpose<Inverse<Matrix>>&& expression);
+      // operators
+      using BLASMatrix<T>::operator=;
+      using BLASMatrix<T>::operator*=;
 
       [[nodiscard]] T& entry(size_t row_index, size_t column_index);
       [[nodiscard]] const T& entry(size_t row_index, size_t column_index) const;
       // vector view
       [[nodiscard]] MutableVectorView<std::vector<double>> column(size_t column_index);
       [[nodiscard]] VectorView<std::vector<double>> column(size_t column_index) const;
+      [[nodiscard]] Submatrix submatrix(size_t number_rows, size_t number_columns);
 
-      [[nodiscard]] T* data();
-      [[nodiscard]] const T* data() const;
+      [[nodiscard]] T* data() override;
+      [[nodiscard]] const T* data() const override;
       void fill(T value);
       void print(std::ostream& stream) const;
-
-      const size_t number_rows, number_columns;
 
    protected:
       std::vector<T> matrix; // column-major ordering
@@ -53,70 +71,8 @@ namespace uno {
 
    template <typename T>
    DenseMatrix<T>::DenseMatrix(size_t number_rows, size_t number_columns):
-         number_rows(number_rows), number_columns(number_columns),
+         BLASMatrix<T>(number_rows, number_columns, number_rows),
          matrix(number_rows * number_columns, T(0)) {
-   }
-
[…]
