### 2026-02-26 dd59af2
Moved computation of Cholesky factors to BLASMatrix
