### 2026-02-26 0ef7986
Use operator overloading to compute low-rank updates C := A A^T and C += alpha A^T A + calls to BLAS
