+++ uno/linear_algebra/BLASMatrix.hpp
@@ -30,6 +30,14 @@ namespace uno {
       template <typename Matrix1, typename Matrix2, typename Matrix3>
       BLASMatrix& operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2, Transpose<Matrix3>>>&& expression);
 
+      // specialized operator+= for low-rank update C := A A^T
+      template <typename Matrix>
+      BLASMatrix& operator=(Multiplication<Matrix, Transpose<Matrix>>&& expression);
+
+      // specialized operator+= for low-rank update C += alpha A^T A
+      template <typename Matrix>
+      BLASMatrix& operator+=(ScalarMultiple<Multiplication<Transpose<Matrix>, Matrix>>&& expression);
+
       // specialized operator= for B := B A⁻ᵀ
       template <typename Matrix>
       BLASMatrix& operator*=(Transpose<Inverse<Matrix>>&& expression);
@@ -63,7 +71,7 @@ namespace uno {
    template <typename Matrix1, typename Matrix2, typename Matrix3>
    BLASMatrix<T>& BLASMatrix<T>::operator=(Sum<ScalarMultiple<Matrix1>, Multiplication<Matrix2,
          Transpose<Matrix3>>>&& expression) {
-      const double beta = expression.get_left().get_factor();
+      const T beta = expression.get_left().get_factor();
       const auto& C = expression.get_left().get_expression();
       const auto& A = expression.get_right().get_left();
       const auto& B = expression.get_right().get_right().get_matrix();
@@ -72,10 +80,10 @@ namespace uno {
          throw std::runtime_error("BLASMatrix::operator= is not defined when C != this");
       }
       if (this->number_rows != A.number_rows || this->number_columns != B.number_rows) {
-         throw std::runtime_error("Dimension mismatch in BLASMatrix::operator=");
+         throw std::runtime_error("Dimension mismatch");
       }
-      const char transa = 'N'; // A
-      const char transb = 'T'; // B^T
+      constexpr char transa = 'N'; // A
+      constexpr char transb = 'T'; // B^T
       const int m = static_cast<int>(A.number_rows); // number of rows of A
       const int n = static_cast<int>(B.number_rows); // number of columns of B^T/number of rows of B
       const int k = static_cast<int>(A.number_columns); // number of columns of A
@@ -88,6 +96,62 @@ namespace uno {
       return *this;
    }
 
+   // specialized operator+= for low-rank update C := A A^T
+   template <typename T>
+   template <typename Matrix>
+   BLASMatrix<T>& BLASMatrix<T>::operator=(Multiplication<Matrix, Transpose<Matrix>>&& expression) {
+      assert(this->number_rows == this->number_columns);
+      // decode expression as alpha A B^T
+      const auto& A = expression.get_left();
+      const auto& B = expression.get_right().get_matrix();
+      const size_t correction_rank = A.number_columns;
+      DEBUG << "Performing rank " << correction_rank << " update\n";
+      // check that A and B are the same object
+      if (&A != &B) {
+         throw std::runtime_error("BLASMatrix::operator+=: low-rank update called on two different correction matrices");
+      }
+      constexpr char uplo = 'L'; // lower triangular
+      constexpr char trans = 'N'; // A A^T
+      const int n = static_cast<int>(this->number_rows); // dimension of matrix
+      const int k = static_cast<int>(correction_rank); // number of columns of A
+      constexpr double alpha = 1.;
+      const int lda = static_cast<int>(A.leading_dimension); // leading dimension of A
+      constexpr double beta = 0.;
+      const int ldc = static_cast<int>(this->leading_dimension); // leading dimension of matrix
[…]
