+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.hpp
@@ -68,11 +68,6 @@ namespace uno {
       void update_D();
       void recompute_hessian_representation();
       [[nodiscard]] double compute_delta() const;
-
-      static void perform_high_rank_update(DenseMatrix<double>& matrix, size_t matrix_dimension, size_t matrix_leading_dimension,
-         DenseMatrix<double>& high_rank_correction, size_t correction_rank, size_t correction_leading_dimension, double alpha, double beta);
-      static void perform_high_rank_update_transpose(DenseMatrix<double>& matrix, size_t matrix_dimension, size_t matrix_leading_dimension,
-         DenseMatrix<double>& high_rank_correction, size_t correction_rank, size_t correction_leading_dimension, double alpha, double beta);
    };
 } // namespace
 
