### 2026-02-26 7de0593
example_hs015.c: added a run with L-BFGS Hessian
