### 2026-02-26 42dd51b
example_hs015.py: added a run with L-BFGS Hessian
