### 2026-02-26 d89093d
NLP_wrapper.jl: added a run with L-BFGS Hessian
