### 2026-02-26 9786f80
Finalize the CUTEstModel in NLP_wrapper.jl
