+++ uno/optimization/Iterate.cpp
@@ -2,8 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
 #include "Iterate.hpp"
-#include "linear_algebra/Vector.hpp"
-#include "model/Model.hpp"
 
 namespace uno {
    Iterate::Iterate(size_t number_variables, size_t number_constraints) :
@@ -27,7 +25,7 @@ namespace uno {
 
       stream << "          ┌ Stationarity: " << iterate.residuals.stationarity << '\n';
       stream << "Residuals │ Complementarity: " << iterate.residuals.complementarity << '\n';
-      stream << "          └ Lagrangian gradient: " << iterate.residuals.lagrangian_gradient;
+      stream << "          └ Lagrangian gradient: " << iterate.residuals.lagrangian_gradient << '\n';
 
       stream << "                  ┌ Infeasibility: " << iterate.progress.infeasibility << '\n';
       stream << "Progress measures │ Optimality: " << iterate.progress.objective(1.) << '\n';
