+++ uno/optimization/OptimizationProblem.hpp
@@ -46,8 +46,8 @@ namespace uno {
       virtual void evaluate_constraints(const Iterate& iterate, double* constraints, Evaluations& evaluations) const;
       virtual void evaluate_objective_gradient(const Iterate& iterate, double* objective_gradient, Evaluations& evaluations) const;
       virtual void evaluate_jacobian(const Vector<double>& primals, double* jacobian_values, Evaluations& evaluations) const;
-      virtual void evaluate_lagrangian_gradient(Vector<double>& lagrangian_gradient, const Iterate& iterate,
-         Evaluations& evaluations) const;
+      virtual void evaluate_lagrangian_gradient(const Iterate& iterate, Evaluations& evaluations,
+         Vector<double>& lagrangian_gradient) const;
       virtual void evaluate_lagrangian_hessian(Statistics& statistics, HessianModel& hessian_model,
          const Vector<double>& primal_variables, const Multipliers& multipliers, double* hessian_values) const;
       virtual void compute_hessian_vector_product(HessianModel& hessian_model, const double* x, const double* vector,
