+++ uno/ingredients/hessian_models/ZeroHessian.hpp
@@ -10,6 +10,7 @@ namespace uno {
    class ZeroHessian : public HessianModel {
    public:
       explicit ZeroHessian(size_t number_variables);
+      ~ZeroHessian() override = default;
 
       [[nodiscard]] bool has_hessian_operator() const override;
       [[nodiscard]] bool has_hessian_matrix() const override;
@@ -18,6 +19,9 @@ namespace uno {
       void compute_sparsity(uno_int* row_indices, uno_int* column_indices, uno_int solver_indexing) const override;
       [[nodiscard]] bool is_positive_definite() const override;
 
+      void initialize_statistics(Statistics& statistics) const override;
+      void notify_accepted_iterate(Statistics& statistics, const Iterate& current_iterate, const Iterate& trial_iterate,
+         EvaluationCache& evaluation_cache) override;
       void evaluate_hessian(Statistics& statistics, const Vector<double>& primal_variables, double objective_multiplier,
          const Vector<double>& constraint_multipliers, double* hessian_values) override;
       void compute_hessian_vector_product(const double* x, const double* vector, double objective_multiplier,

[…]
