### 2026-02-26 a57290f
[WIP] L-BFGS approximation for SQP methods (#512)
* Copied content of PR #150

* Moved evaluate_lagrangian_gradient to Model, with multipliers and objective_multiplier as parameters. Adapted the calls in LBFGSHessian

* Added Julia tests with L-BFGS Hessian

* Enable LAPACK in build_tarballs_*.jl

* Used a LS method instead of TR in the L-BFGS Julia tests

* Forgot to update Action

* Excluded some Julia tests, mostly for good reasons (hard to detect unboundedness with positive definite Hessian) except for hs109

* Renamed L-BFGS into LBFGS, with default value 6 (a la IPOPT), and documented it
[…]
