+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.cpp
@@ -1,0 +1,235 @@
+// Copyright (c) 2025-2026 Charlie Vanaret
+// Licensed under the MIT license. See LICENSE file in the project directory for details.
+
+#include "LBFGSHessian.hpp"
+#include "model/Model.hpp"
+#include "options/Options.hpp"
+#include "linear_algebra/BLAS.hpp"
+#include "linear_algebra/Vector.hpp"
+#include "linear_algebra/VectorView.hpp"
+#include "optimization/EvaluationCache.hpp"
+#include "optimization/Iterate.hpp"
+#include "symbolic/Inverse.hpp"
+#include "symbolic/Multiplication.hpp"
+#include "symbolic/Range.hpp"
+#include "symbolic/ScalarMultiple.hpp"
+#include "symbolic/Subtraction.hpp"
+#include "symbolic/Sum.hpp"
+#include "symbolic/Transpose.hpp"
+#include "tools/Logger.hpp"
+#include "tools/Statistics.hpp"
+
+namespace uno {
+   LBFGSHessian::LBFGSHessian(const Model& model, double objective_multiplier, const Options& options):
+         HessianModel("L-BFGS"),
+         model(model),
+         fixed_objective_multiplier(objective_multiplier),
+         memory_size(options.get_unsigned_int("quasi_newton_memory_size")),
+         // matrices
+         S(this->model.number_variables, this->memory_size),
+         Y(this->model.number_variables, this->memory_size),
+         L(this->memory_size, this->memory_size),
+         D(this->memory_size),
+         invsqrt_D(this->memory_size),
+         L_invsqrt_D(this->memory_size, this->memory_size),
+         M(this->memory_size, this->memory_size),
+         U(this->model.number_variables, this->memory_size),
+         V(this->model.number_variables, this->memory_size),
+         current_lagrangian_gradient(this->model.number_variables),
+         trial_lagrangian_gradient(this->model.number_variables) {
+      if (this->memory_size <= 0) {
+         throw std::runtime_error("The quasi-Newton memory size should be positive");
+      }
+   }
+
+   bool LBFGSHessian::has_hessian_operator() const {
+      return true;
+   }
+
+   bool LBFGSHessian::has_hessian_matrix() const {
+      // never form the explicit, dense matrix
+      return false;
+   }
+
+   bool LBFGSHessian::has_curvature() const {
+      return true;
+   }
+
+   size_t LBFGSHessian::number_nonzeros() const {
+      throw std::runtime_error("LBFGSHessian::number_nonzeros should not be called");
+   }
+
+   void LBFGSHessian::compute_sparsity(int* /*row_indices*/, int* /*column_indices*/, int /*solver_indexing*/) const {
+      throw std::runtime_error("LBFGSHessian::compute_sparsity should not be called");
+   }
+
+   bool LBFGSHessian::is_positive_definite() const {
+      return true;
+   }
+
+   void LBFGSHessian::initialize_statistics(Statistics& statistics) const {
+      statistics.add_column("|BFGS|", Statistics::double_width - 2, 2, Statistics::column_order.at("|BFGS|"));
+      statistics.set("|BFGS|", this->number_entries_in_memory);
+   }
+
+   void LBFGSHessian::notify_accepted_iterate(Statistics& statistics, const Iterate& current_iterate, const Iterate& trial_iterate,
+         EvaluationCache& evaluation_cache) {
+      DEBUG << "\n*** Updating the BFGS memory at slot " << this->current_index << '\n';
+      // update the matrices S, Y and D
+      this->update_S(current_iterate, trial_iterate);
+      this->update_Y(current_iterate, trial_iterate, evaluation_cache);
[…]
