+++ uno/ingredients/hessian_models/ExactHessian.cpp
@@ -33,6 +33,13 @@ namespace uno {
       return false;
    }
 
+   void ExactHessian::initialize_statistics(Statistics& /*statistics*/) const {
+   }
+
+   void ExactHessian::notify_accepted_iterate(Statistics& /*statistics*/, const Iterate& /*current_iterate*/,
+         const Iterate& /*trial_iterate*/, EvaluationCache& /*evaluation_cache*/) {
+   }
+
    void ExactHessian::evaluate_hessian(Statistics& /*statistics*/, const Vector<double>& primal_variables,
          double objective_multiplier, const Vector<double>& constraint_multipliers, double* hessian_values) {
       this->model.evaluate_lagrangian_hessian(primal_variables, objective_multiplier, constraint_multipliers, hessian_values);
