+++ docs/developers/checklists.md
@@ -11,10 +11,10 @@ In preparation for an Uno release, use the following checklist. The PR should ha
 ````
 ## Pre-release
 
-- [ ] Change the version number in `Uno::current_version()` in `uno/Uno.cpp`
 - [ ] Change the version number in `CITATION.cff` (twice)
 - [ ] Change the version number in `CMakeLists.txt`
 - [ ] Change the version number in `interfaces/C/Uno_C_API.h`
+- [ ] Change the version number in `interfaces/Julia/Project.toml`
 - [ ] Update the logo in `docs/figures/logo.png`
 - [ ] The commit messages in this PR do not contain `[ci skip]`
 
