+++ docs/developers/checklists.md
@@ -14,14 +14,15 @@ In preparation for an Uno release, use the following checklist. The PR should ha
 - [ ] Change the version number in `CITATION.cff` (twice)
 - [ ] Change the version number in `CMakeLists.txt`
 - [ ] Change the version number in `interfaces/C/Uno_C_API.h`
-- [ ] Change the version number in `interfaces/Julia/Project.toml`
+- [ ] Change the version number in `interfaces/Fortran/uno_c.f90`
 - [ ] Update the logo in `docs/figures/logo.png`
 - [ ] The commit messages in this PR do not contain `[ci skip]`
 
 ## Post-release
 
-- `Uno_jll.jl`
+- `Uno_jll.jl` and `UnoSolver.jl`
   - [ ] Update the [Yggdrasil tarballs](https://github.com/JuliaPackaging/Yggdrasil/blob/master/U/Uno/build_tarballs.jl)
+  - [ ] Change the patch version number in `interfaces/Julia/Project.toml`
   - [ ] Change the `Uno_jll` version number in `interfaces/Julia/Project.toml`
   - [ ] Change the `Uno_jll` version number in `interfaces/Julia/gen/Project.toml`
 - [ ] Update the logo in [the Github settings](https://github.com/cvanaret/Uno/settings)
