+++ docs/developers/checklists.md
@@ -22,11 +22,10 @@ In preparation for an Uno release, use the following checklist. The PR should ha
 
 - `Uno_jll.jl` and `UnoSolver.jl`
   - [ ] Update the [Yggdrasil tarballs](https://github.com/JuliaPackaging/Yggdrasil/blob/master/U/Uno/build_tarballs.jl)
-  - [ ] Change the patch version number in `interfaces/Julia/Project.toml`
   - [ ] Change the `Uno_jll` version number in `interfaces/Julia/Project.toml`
   - [ ] Change the `Uno_jll` version number in `interfaces/Julia/gen/Project.toml`
+  - [ ] Change the patch version number in `interfaces/Julia/Project.toml`
 - [ ] Update the logo in [the Github settings](https://github.com/cvanaret/Uno/settings)
-- [ ] Update the logo in `docs/figures/logo.png`
 ````
 
 ## Making a release of `UnoSolver.jl`
