+++ interfaces/Fortran/uno_c.f90
@@ -77,8 +77,8 @@ integer(uno_int), parameter :: UNO_UNBOUNDED                   = 6
 ! Uno version
 !---------------------------------------------
 integer(uno_int), parameter :: UNO_VERSION_MAJOR = 2
-integer(uno_int), parameter :: UNO_VERSION_MINOR = 3
-integer(uno_int), parameter :: UNO_VERSION_PATCH = 1
+integer(uno_int), parameter :: UNO_VERSION_MINOR = 4
+integer(uno_int), parameter :: UNO_VERSION_PATCH = 0
 
 !---------------------------------------------
 ! uno_get_version
