### 2026-02-27 b9c876c
Prep for v2.4.0 (#532)
