### 2026-02-27 bdce819
[UnoSolver.jl] Release 0.2.2 (#533)
