+++ pyproject.toml
@@ -7,10 +7,7 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "unopy"
-# PyPi
-#version = "0.1.2"
-# TestPyPi
-version = "0.2.0"
+version = "0.2.1"
 #requires-python = ">=3.8"
 description = "Uno, a highly customizable SQP & barrier solver for nonlinearly constrained optimization"
 readme = "interfaces/Python/README.md"
