### 2026-02-27 620a644
[unopy] Release v0.2.1 (#535)
Minor change
