+++ uno/ingredients/subproblem_solvers/BoundConstrainedSolver.hpp
@@ -22,10 +22,11 @@ namespace uno {
 
       void initialize_memory(const Subproblem& subproblem) override = 0;
 
-      void solve(Statistics& statistics, Subproblem& subproblem, double trust_region_radius, const Vector<double>& initial_point,
-         Direction& direction, const WarmstartInformation& warmstart_information) override = 0;
+      void solve(Statistics& statistics, Subproblem& subproblem, double trust_region_radius,
+         const Vector<double>& initial_point, Direction& direction, Evaluations& current_evaluations,
+         const WarmstartInformation& warmstart_information) override = 0;
 
-      [[nodiscard]] virtual EvaluationSpace& get_evaluation_space() = 0;
+      [[nodiscard]] virtual SolverWorkspace& get_evaluation_space() = 0;
    };
 } // namespace
 
