+++ uno/ingredients/subproblem_solvers/dogleg/DoglegWorkspace.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2025 Charlie Vanaret
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
-#include "DoglegEvaluationSpace.hpp"
+#include "DoglegWorkspace.hpp"
 #include "ingredients/subproblem/Subproblem.hpp"
 #include "ingredients/subproblem_solvers/SymmetricIndefiniteLinearSolver.hpp"
 #include "optimization/Direction.hpp"
@@ -9,7 +9,7 @@
 #include "optimization/WarmstartInformation.hpp"
 
 namespace uno {
-   void DoglegEvaluationSpace::initialize_memory(const Subproblem& subproblem) {
+   void DoglegWorkspace::initialize_memory(const Subproblem& subproblem) {
       // Newton step
       this->objective_gradient.resize(subproblem.number_variables);
       this->newton_step.resize(subproblem.number_variables);
@@ -18,42 +18,25 @@ namespace uno {
       this->cauchy_step.resize(subproblem.number_variables);
    }
 
-   void DoglegEvaluationSpace::evaluate_constraint_jacobian(const OptimizationProblem& /*problem*/, Iterate& /*iterate*/) {
-      // do nothing
-   }
-
-   void DoglegEvaluationSpace::compute_constraint_jacobian_vector_product(const Vector<double>& /*vector*/, Vector<double>& result) const {
-      result.fill(0.);
-   }
-
-   void DoglegEvaluationSpace::compute_constraint_jacobian_transposed_vector_product(const Vector<double>& /*vector*/,
-         Vector<double>& result) const {
-      result.fill(0.);
-   }
-
-   double DoglegEvaluationSpace::compute_hessian_quadratic_product(const Subproblem& /*subproblem*/,
+   double DoglegWorkspace::compute_hessian_quadratic_product(const Subproblem& /*subproblem*/,
          const Vector<double>& /*vector*/) const {
       throw std::runtime_error("Not implemented yet");
    }
 
-   void DoglegEvaluationSpace::evaluate_objective_gradient(const Subproblem& subproblem, const WarmstartInformation& warmstart_information) {
-      if (warmstart_information.objective_changed) {
-         subproblem.problem.evaluate_objective_gradient(subproblem.current_iterate, this->objective_gradient.data());
-      }
-   }
-
-   void DoglegEvaluationSpace::compute_newton_step(Statistics& statistics, const Subproblem& subproblem,
-         SymmetricIndefiniteLinearSolver<double>& linear_solver, Direction& direction, const WarmstartInformation& warmstart_information) {
-      if (warmstart_information.objective_changed) {
+   void DoglegWorkspace::compute_newton_step(Statistics& statistics, const Subproblem& subproblem, Direction& direction,
+         SymmetricIndefiniteLinearSolver<double>& linear_solver, Evaluations& current_evaluations,
+         const WarmstartInformation& warmstart_information) {
+      if (warmstart_information.new_iterate) {
          // g = ∇f(x_k)
-         subproblem.problem.evaluate_objective_gradient(subproblem.current_iterate, this->objective_gradient.data());
-         linear_solver.solve_indefinite_system(statistics, subproblem, direction, warmstart_information);
+         subproblem.problem.evaluate_objective_gradient(subproblem.current_iterate, this->objective_gradient.data(),
+            current_evaluations);
+         linear_solver.solve_indefinite_system(statistics, subproblem, direction, current_evaluations, warmstart_information);
          this->newton_step = direction.primals;
          this->newton_step_squared_norm = dot(this->newton_step, this->newton_step);
       }
    }
 
[…]
