+++ uno/ingredients/subproblem_solvers/dogleg/DoglegWorkspace.hpp
@@ -1,23 +1,26 @@
 // Copyright (c) 2025 Charlie Vanaret
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
-#ifndef UNO_DOGLEGEVALUATIONSPACE_H
-#define UNO_DOGLEGEVALUATIONSPACE_H
+#ifndef UNO_DOGLEGWORKSPACE_H
+#define UNO_DOGLEGWORKSPACE_H
 
-#include "optimization/EvaluationSpace.hpp"
+#include "../SolverWorkspace.hpp"
 #include "linear_algebra/Vector.hpp"
 #include "tools/Infinity.hpp"
 
 namespace uno {
    // forward declaration
    class Direction;
+   class Evaluations;
+   class Statistics;
    template <typename ElementType>
    class SymmetricIndefiniteLinearSolver;
+   class WarmstartInformation;
 
-   class DoglegEvaluationSpace: public EvaluationSpace {
+   class DoglegWorkspace: public SolverWorkspace {
    public:
-      DoglegEvaluationSpace() = default;
-      ~DoglegEvaluationSpace() override = default;
+      DoglegWorkspace() = default;
+      ~DoglegWorkspace() override = default;
 
       // Newton step
       Vector<double> objective_gradient{};
@@ -31,16 +34,13 @@ namespace uno {
 
       void initialize_memory(const Subproblem& subproblem);
 
-      void evaluate_constraint_jacobian(const OptimizationProblem& /*problem*/, Iterate& /*iterate*/) override;
-      void compute_constraint_jacobian_vector_product(const Vector<double>& /*vector*/, Vector<double>& result) const override;
-      void compute_constraint_jacobian_transposed_vector_product(const Vector<double>& vector, Vector<double>& result) const override;
       [[nodiscard]] double compute_hessian_quadratic_product(const Subproblem& subproblem,
          const Vector<double>& vector) const override;
 
-      void evaluate_objective_gradient(const Subproblem& subproblem, const WarmstartInformation& warmstart_information);
-      void compute_newton_step(Statistics& statistics, const Subproblem& subproblem,
-         SymmetricIndefiniteLinearSolver<double>& linear_solver, Direction& direction, const WarmstartInformation& warmstart_information);
-      void compute_dogleg(const Subproblem& subproblem, Direction& direction, const WarmstartInformation& warmstart_information);
+      void compute_newton_step(Statistics& statistics, const Subproblem& subproblem, Direction& direction,
+         SymmetricIndefiniteLinearSolver<double>& linear_solver, Evaluations& current_evaluations, const WarmstartInformation& warmstart_information);
+      void compute_dogleg(const Subproblem& subproblem, Direction& direction, Evaluations& current_evaluations,
+         const WarmstartInformation& warmstart_information);
 
    private:
       void compute_cauchy_step(const Subproblem& subproblem, const WarmstartInformation& warmstart_information);
@@ -48,4 +48,4 @@ namespace uno {
    };
 } // namespace
 
-#endif // UNO_DOGLEGEVALUATIONSPACE_H
+#endif // UNO_DOGLEGWORKSPACE_H
