### 2026-02-27 313a5d4
[UnoSolver.jl] Pass the preset before the other options
