### 2026-02-27 bcab160
Apply suggestion from @amontoison
