### 2026-02-27 7276f0e
[UnoSolver.jl] Pass the preset before the other options
