### 2026-02-28 fd923bc
[UnoSolver.jl] Pass the preset before the other options (#538)
* [UnoSolver.jl] Pass the preset before the other options

* Apply suggestion from @amontoison

* [UnoSolver.jl] Pass the preset before the other options
