+++ uno/ingredients/subproblem_solvers/SSIDS/SSIDSSolver.cpp
@@ -1,0 +1,101 @@
+#include <iostream>
+#include "SSIDSSolver.hpp"
+#include "ingredients/subproblem/Subproblem.hpp"
+#include "optimization/Direction.hpp"
+
+//export OMP_PROC_BIND=true
+//export OMP_CANCELLATION=true
+
+namespace uno {
+   SSIDSSolver::SSIDSSolver(): DirectSymmetricIndefiniteLinearSolver<double>() {
+      spral_ssids_default_options(&this->workspace.options);
+      this->workspace.options.array_base = 1; // Fortran indexing
+      this->workspace.options.print_level = -1; // no printing
+   }
+
+   void SSIDSSolver::initialize_hessian(const Subproblem& subproblem) {
+      this->coo_workspace.initialize_hessian(subproblem);
+
+      // workspace
+      const size_t dimension = subproblem.number_variables;
+      this->workspace.n = static_cast<int>(dimension);
+      this->workspace.nnz = static_cast<int>(this->coo_workspace.number_matrix_nonzeros);
+   }
+
+   void SSIDSSolver::initialize_augmented_system(const Subproblem& subproblem) {
+      this->coo_workspace.initialize_augmented_system(subproblem);
+
+      // workspace
+      const size_t dimension = subproblem.number_variables + subproblem.number_constraints;
+      this->workspace.n = static_cast<int>(dimension);
+      this->workspace.nnz = static_cast<int>(this->coo_workspace.number_matrix_nonzeros);
+   }
+
+   void SSIDSSolver::do_symbolic_analysis() {
+      spral_ssids_analyse_coord(this->workspace.n, nullptr, this->workspace.nnz, this->coo_workspace.matrix_row_indices.data(),
+         this->coo_workspace.matrix_column_indices.data(), nullptr, &this->workspace.akeep, &this->workspace.options,
+         &this->workspace.inform);
+      if (this->workspace.inform.flag < 0) {
+         std::cout << "ERROR in SSIDSSolver::do_symbolic_analysis\n";
+         spral_ssids_free(&this->workspace.akeep, &this->workspace.fkeep);
+      }
+   }
+
+   void SSIDSSolver::do_numerical_factorization(const double* matrix_values, bool is_matrix_positive_definite) {
+      spral_ssids_factor(is_matrix_positive_definite, nullptr, nullptr, matrix_values, nullptr, this->workspace.akeep,
+         &this->workspace.fkeep, &this->workspace.options, &this->workspace.inform);
+      if(this->workspace.inform.flag < 0) {
+         std::cout << "ERROR in SSIDSSolver::do_numerical_factorization\n";
+         spral_ssids_free(&this->workspace.akeep, &this->workspace.fkeep);
+      }
+   }
+
+   void SSIDSSolver::solve_indefinite_system(const Vector<double>& /*matrix_values*/, const Vector<double>& rhs, Vector<double>& result) {
+      result = rhs;
+      spral_ssids_solve1(0, result.data(), this->workspace.akeep, this->workspace.fkeep, &this->workspace.options,
+         &this->workspace.inform);
+      if (this->workspace.inform.flag < 0) {
+         std::cout << "ERROR in SSIDSSolver::solve_indefinite_system\n";
+         spral_ssids_free(&this->workspace.akeep, &this->workspace.fkeep);
+      }
+   }
+
+   void SSIDSSolver::solve_indefinite_system(Statistics& statistics, const Subproblem& subproblem, Direction& direction,
+         Evaluations& current_evaluations, const WarmstartInformation& warmstart_information) {
+      // set up the linear system by evaluating the functions at the current iterate
[…]
