+++ CMakeLists.txt
@@ -206,6 +206,14 @@ if(MUMPS_LIBRARY)
    message(STATUS "Found MUMPS")
 endif()
 
+# SPRAL SSIDS
+find_library(SPRAL NAMES spral)
+if(SPRAL)
+   list(APPEND UNO_SOURCE_FILES
+           uno/ingredients/subproblem_solvers/SSIDS/SSIDSSolver.cpp)
+   link_to_uno(spral ${SPRAL})
+endif()
+
 # LAPACK
 # convert relative paths of LAPACK to absolute paths
 if(LAPACK_LIBRARIES)
