### 2026-03-01 4d92ee2
Workflows: ignore if documentation files are changed
