+++ .github/workflows/unit-tests.yml
@@ -7,12 +7,16 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.ya?ml'
+      - 'docs/**'
   pull_request:
     branches: [ "main" ]
     paths-ignore:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.ya?ml'
+      - 'docs/**'
 
 jobs:
   build:
