+++ .github/workflows/build-python-wheels.yml
@@ -8,12 +8,16 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.ya?ml'
+      - 'docs/**'
   pull_request:
     branches: [ "main" ]
     paths-ignore:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.ya?ml'
+      - 'docs/**'
 
 jobs:
   build_wheels:
