+++ docs/index.md
@@ -1,6 +1,2 @@
-Welcome to Uno's documentation
-==============================
-
-See the available configuration options on the `Options` page:
-
-- Options: `options.md`
+Uno, a unified solver for nonlinearly constrained optimization
+==============================================================
