+++ mkdocs.yml
@@ -3,5 +3,5 @@ theme:
   name: material
 nav:
   - Introduction: 'index.md'
-  - Options: `options.md`
+  - Options: 'options.md'
   - 'Issue Tracker': 'https://github.com/cvanaret/Uno/issues'
