+++ .github/workflows/julia-tests-ubuntu.yml
@@ -6,7 +6,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
   pull_request:
     branches: [main]
@@ -14,7 +15,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
     
 # needed to allow julia-actions/cache to delete old caches that it has created
