+++ .github/workflows/build-python-wheels.yml
@@ -8,7 +8,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
   pull_request:
     branches: [ "main" ]
@@ -16,7 +17,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
 
 jobs:
