+++ .github/workflows/build.yml
@@ -7,7 +7,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
   pull_request:
     branches: [ "main" ]
@@ -15,7 +16,8 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
-      - '*.ya?ml'
+      - '*.yml'
+      - '*.yaml'
       - 'docs/**'
 
 jobs:
