+++ mkdocs.yml
@@ -5,3 +5,4 @@ nav:
   - Introduction: 'index.md'
   - Options: 'options.md'
   - 'Issue Tracker': 'https://github.com/cvanaret/Uno/issues'
+  - Acknowledgments: 'acknowledgments.md'
