+++ mkdocs.yml
@@ -3,6 +3,9 @@ theme:
   name: material
 nav:
   - Introduction: 'index.md'
-  - Options: 'options.md'
-  - 'Issue Tracker': 'https://github.com/cvanaret/Uno/issues'
+  - 'Unification framework': 'unification_framework.md'
+  - 'Uno options': 'options.md'
+  - 'Issue tracker': 'https://github.com/cvanaret/Uno/issues'
+  - 'Checklists for developers': 'developers/checklists.md'
   - Acknowledgments: 'acknowledgments.md'
+  
