### 2026-03-01 3ecc2c5
Added support for MathJax
