### 2026-03-01 6483778
Trying optimization problem description with LaTeX array
