### 2026-03-01 fcee383
Fixed optimization problems
