+++ docs/unification_framework.md
@@ -79,6 +79,169 @@ The inner loop (**repeat**) generates and solves a feasible subproblem (possibly
 
 The "wheel of strategies" (Figure 1) organizes a wide range of strategies into a coherent hierarchy. The outer layer represents the optimization layers, the middle layer represents the ingredients, and the inner layer represents the strategies. Strategies that perform similar tasks within an optimization method are listed under the same ingredient (e.g., a line search and a trust-region method are both globalization mechanisms).
 
-![Unifying framework: the "wheel of strategies".](figures/wheel.png)
+<p align="center">
+   <img src="figures/wheel.png" alt="Unifying framework: the "wheel of strategies"." width="60%" /> \\
+   **Figure 1.** Unifying framework: the "wheel of strategies".
+</p>
 
-**Figure 1.** Unifying framework: the "wheel of strategies".
+## Unified view of state-of-the-art techniques
+
+In this section we describe a set of strategies that fall into each of the following ingredients of our unifying framework: globalization strategies, constraint relaxation strategies, inequality handling methods, Hessian models, inertia correction strategies, and globalization mechanisms. Our goal is to illustrate the wide variety of methods within a common notation to motivate the design of our modular solver.
+
+### Globalization strategies
+
+Constrained optimization methods must achieve two competing goals: minimizing the objective function and minimizing the constraint violation. Globalization strategies determine whether a trial iterate $\hat{x}^{(k+1)} \stackrel{\text{def}}{=} x^{(k)} + \alpha d_x^{(k)}$ (for a given step length $\alpha \in (0, 1]$) makes acceptable progress with respect to these goals. We consider strategies that ensure *global convergence*, that is, convergence to a local minimum, or (weaker) stationary point, from any starting point.
+
+Three (possibly primal-dual) *progress measures* are monitored throughout the optimization process:
+1. an **infeasibility measure** $\eta$ (typically $\|c(x)\|$ for some norm);
+2. an **objective measure** $\omega_\pi$ parameterized by the objective multiplier $\pi \ge 0$ (typically $\pi f(x)$); and
+3. an **auxiliary measure** $\xi$ (terms such as barrier and proximal terms).
+
+The local models of $\eta(x)$, $\omega_\pi(x)$, and $\xi(x)$ at iteration $k$ about the current iterate are denoted by $\mathbf{\eta}^{(k)}(d_x)$, $\mathbf{\omega}_\pi^{(k)}(d_x)$, and $\mathbf{\xi}^{(k)}(d_x)$ for a given primal direction $d_x$. We define the respective *predicted reductions*:
+
+$$\Delta \mathbf{\eta}^{(k)}(d_x) \stackrel{\text{def}}{=} \mathbf{\eta}^{(k)}(0) - \mathbf{\eta}^{(k)}(d_x)$$
+
+$$\Delta \mathbf{\omega}_\pi^{(k)}(d_x) \stackrel{\text{def}}{=} \mathbf{\omega}_\pi^{(k)}(0) - \mathbf{\omega}_\pi^{(k)}(d_x)$$
+
+$$\Delta \mathbf{\xi}^{(k)}(d_x) \stackrel{\text{def}}{=} \mathbf{\xi}^{(k)}(0) - \mathbf{\xi}^{(k)}(d_x)$$
+
+In order to ensure convergence, the progress measures and their local models must be intimately linked to the definitions of the problem and the subproblem.
+
[…]
