### 2026-03-01 0d3f825
Fixed missing \end{array} and replaced PDF image with PNG
