### 2026-03-01 2642701
Added support for footnotes
