### 2026-03-01 606efab
Added section Uno solver
