+++ docs/uno_solver.md
@@ -1,0 +1,88 @@
+## Uno: a unified solver for nonlinearly constrained optimization
+
+We have implemented our unifying framework for nonlinearly constrained optimization within Uno, a modular solver written in C++17. A generic and flexible code, it supports a broad range of strategies that can be combined automatically and on the fly with no programming effort from the user. The code is packaged in a lightweight library (around 10,000 lines of code for the current version, excluding interfaces) available as open-source software under the MIT license at https://github.com/cvanaret/Uno. Uno is available via its C, Julia (registered package `UnoSolver.jl`), Python (the packaging of `unopy` is under development), Fortran, and AMPL interfaces.
+
+### Generic architecture
+
+The modularity of Uno stems from its generic architecture: each ingredient is implemented once and independently of the others, which improves readability, and makes building blocks less prone to error and easier to maintain than monolithic codes.
+
+Figure 3 represents Uno's object-oriented architecture as a Unified Modeling Language (UML) diagram based on inheritance ("is a") and composition ("has a"). The ingredients are modeled as abstract classes: they define interfaces, that is, generic actions and behaviors that must be implemented by concrete strategies modeled as subclasses. For example:
+- the classes `BacktrackingLineSearch` and `TrustRegionStrategy` inherit from the abstract class `GlobalizationMechanism` and thus must implement the purely virtual member function `compute_acceptable_iterate()`;
+- the `GlobalizationMechanism` class possesses a member of type `ConstraintRelaxationStrategy`.
+
+![Uno's UML diagram.](figures/uml_diagram.png)
+
+**Figure 3.** Uno's UML diagram.
+
+### Automatic strategy combinations
+
+Uno implements state-of-the-art strategies that can be combined automatically thanks to the modular software architecture. The number of possible strategy combinations is the size of the Cartesian product of the eight ingredients. Note that all combinations do not necessarily result in sensible algorithms, or even convergent approaches.
+
+At the moment, Uno prohibits the combination of interior-point methods and trust-region strategies. A possible strategy is KNITRO's step decomposition: the direction is decomposed into a normal step that minimizes the constraint violation within the trust region, and a tangential step that minimizes the objective for a given constraint violation target within the trust region. This limitation will be resolved in later Uno versions.
+
+Some strategy combinations are available as "presets" that automatically connect the eight ingredients and set values for the hyperparameters. The following presets are available:
+- **`filtersqp`**: A trust-region restoration filter SQP method à la filterSQP. Second-order correction steps were not implemented.
+- **`ipopt`**: A line-search restoration filter interior-point method à la IPOPT. Second-order correction steps, scaling, least-square multipliers, iterative refinement, iterative bound relaxations, non-monotone techniques, and soft feasibility restoration were not implemented.
+
+### Features of Uno
+
+#### Interfaces to subproblem solvers
+
+Interfaces to the following subproblem solvers are available:
+- **BQPD**: a null-space active-set solver for nonconvex QPs. BQPD accepts Hessian-vector products instead of an explicit matrix;
+- **MA57**, **MA27**, and **MUMPS**: direct solvers for sparse symmetric indefinite linear systems;
+- **HiGHS**: a parallel simplex implementation for linear programming.
+
+#### Definition of the subproblem
+
+The constraint relaxation strategy and the inequality handling method successively reformulate the original problem with respect to the general constraints $c(x) = 0$ and the bound constraints $x \ge 0$. A local model of the resulting reformulated problem is then built.
+
[…]
