### 2026-03-02 ba081a8
Added footer and link to Github
