### 2026-03-02 05ade46
Fixed syntax in mkdocs.yml
