### 2026-03-02 5bf64b8
Fixed opt problem in presets.md
