### 2026-03-02 095b1f5
Fixed equations in presets.md
