### 2026-03-02 fcde857
Removed mention to numerical results
