+++ docs/uno_solver.md
@@ -1,8 +1,8 @@
-## Uno: a unified solver for nonlinearly constrained optimization
+# Uno: a unified solver for nonlinearly constrained optimization
 
 We have implemented our unifying framework for nonlinearly constrained optimization within Uno, a modular solver written in C++17. A generic and flexible code, it supports a broad range of strategies that can be combined automatically and on the fly with no programming effort from the user. The code is packaged in a lightweight library (around 10,000 lines of code for the current version, excluding interfaces) available as open-source software under the MIT license at https://github.com/cvanaret/Uno. Uno is available via its C, Julia (registered package `UnoSolver.jl`), Python (the packaging of `unopy` is under development), Fortran, and AMPL interfaces.
 
-### Generic architecture
+## Generic architecture
 
 The modularity of Uno stems from its generic architecture: each ingredient is implemented once and independently of the others, which improves readability, and makes building blocks less prone to error and easier to maintain than monolithic codes.
 
@@ -14,7 +14,7 @@ Figure 3 represents Uno's object-oriented architecture as a Unified Modeling Lan
 
 **Figure 3.** Uno's UML diagram.
 
-### Automatic strategy combinations
+## Automatic strategy combinations
 
 Uno implements state-of-the-art strategies that can be combined automatically thanks to the modular software architecture. The number of possible strategy combinations is the size of the Cartesian product of the eight ingredients. Note that all combinations do not necessarily result in sensible algorithms, or even convergent approaches.
 
@@ -24,16 +24,16 @@ Some strategy combinations are available as "presets" that automatically connect
 - **`filtersqp`**: A trust-region restoration filter SQP method à la filterSQP. Second-order correction steps were not implemented.
 - **`ipopt`**: A line-search restoration filter interior-point method à la IPOPT. Second-order correction steps, scaling, least-square multipliers, iterative refinement, iterative bound relaxations, non-monotone techniques, and soft feasibility restoration were not implemented.
 
-### Features of Uno
+## Features of Uno
 
-#### Interfaces to subproblem solvers
+### Interfaces to subproblem solvers
 
 Interfaces to the following subproblem solvers are available:
 - **BQPD**: a null-space active-set solver for nonconvex QPs. BQPD accepts Hessian-vector products instead of an explicit matrix;
 - **MA57**, **MA27**, and **MUMPS**: direct solvers for sparse symmetric indefinite linear systems;
 - **HiGHS**: a parallel simplex implementation for linear programming.
 
-#### Definition of the subproblem
+### Definition of the subproblem
 
 The constraint relaxation strategy and the inequality handling method successively reformulate the original problem with respect to the general constraints $c(x) = 0$ and the bound constraints $x \ge 0$. A local model of the resulting reformulated problem is then built.
 
@@ -45,35 +45,35 @@ The subproblem is represented schematically in Figure 4.
 
 **Figure 4.** Definition of a `Subproblem`.
 
-#### Definition of progress measures
+### Definition of progress measures
 
 The progress measures are defined automatically by the successive reformulations of the optimization problem (`l1RelaxedProblem` and `PrimalDualInteriorPointProblem`):
 - the norm of the infeasibility measure is determined by that of the feasibility problem or that of the relaxed problem;
 - the objective measure is $\omega_\pi(x) \stackrel{\text{def}}{=} \pi f(x)$ (for all the currently implemented strategies);
 - the auxiliary measure accumulates proximal (in the feasibility problem `l1RelaxedProblem`) and barrier (in `PrimalDualInteriorPointProblem`) terms.
 
-#### Inertia correction
+### Inertia correction
 
[…]
