+++ docs/uno_solver.md
@@ -7,6 +7,7 @@ We have implemented our unifying framework for nonlinearly constrained optimizat
 The modularity of Uno stems from its generic architecture: each ingredient is implemented once and independently of the others, which improves readability, and makes building blocks less prone to error and easier to maintain than monolithic codes.
 
 Figure 3 represents Uno's object-oriented architecture as a Unified Modeling Language (UML) diagram based on inheritance ("is a") and composition ("has a"). The ingredients are modeled as abstract classes: they define interfaces, that is, generic actions and behaviors that must be implemented by concrete strategies modeled as subclasses. For example:
+
 - the classes `BacktrackingLineSearch` and `TrustRegionStrategy` inherit from the abstract class `GlobalizationMechanism` and thus must implement the purely virtual member function `compute_acceptable_iterate()`;
 - the `GlobalizationMechanism` class possesses a member of type `ConstraintRelaxationStrategy`.
 
@@ -21,6 +22,7 @@ Uno implements state-of-the-art strategies that can be combined automatically th
 At the moment, Uno prohibits the combination of interior-point methods and trust-region strategies. A possible strategy is KNITRO's step decomposition: the direction is decomposed into a normal step that minimizes the constraint violation within the trust region, and a tangential step that minimizes the objective for a given constraint violation target within the trust region. This limitation will be resolved in later Uno versions.
 
 Some strategy combinations are available as "presets" that automatically connect the eight ingredients and set values for the hyperparameters. The following presets are available:
+
 - **`filtersqp`**: A trust-region restoration filter SQP method à la filterSQP. Second-order correction steps were not implemented.
 - **`ipopt`**: A line-search restoration filter interior-point method à la IPOPT. Second-order correction steps, scaling, least-square multipliers, iterative refinement, iterative bound relaxations, non-monotone techniques, and soft feasibility restoration were not implemented.
 
@@ -29,6 +31,7 @@ Some strategy combinations are available as "presets" that automatically connect
 ### Interfaces to subproblem solvers
 
 Interfaces to the following subproblem solvers are available:
+
 - **BQPD**: a null-space active-set solver for nonconvex QPs. BQPD accepts Hessian-vector products instead of an explicit matrix;
 - **MA57**, **MA27**, and **MUMPS**: direct solvers for sparse symmetric indefinite linear systems;
 - **HiGHS**: a parallel simplex implementation for linear programming.
@@ -48,6 +51,7 @@ The subproblem is represented schematically in Figure 4.
 ### Definition of progress measures
 
 The progress measures are defined automatically by the successive reformulations of the optimization problem (`l1RelaxedProblem` and `PrimalDualInteriorPointProblem`):
+
 - the norm of the infeasibility measure is determined by that of the feasibility problem or that of the relaxed problem;
 - the objective measure is $\omega_\pi(x) \stackrel{\text{def}}{=} \pi f(x)$ (for all the currently implemented strategies);
 - the auxiliary measure accumulates proximal (in the feasibility problem `l1RelaxedProblem`) and barrier (in `PrimalDualInteriorPointProblem`) terms.
@@ -55,6 +59,7 @@ The progress measures are defined automatically by the successive reformulations
 ### Inertia correction
 
 Each Hessian model declares whether it is positive definite, and whether it is available as an explicit matrix and as a linear operator. The inertia of the Lagrangian Hessian is corrected if:
+
 - the Hessian model is not already positive definite; and
 - the Hessian model is available as an explicit matrix; and
 - the inertia correction strategy performs primal correction ("primal" or "primal-dual").
[…]
