### 2026-03-02 c2bc683
Basic Readthedocs documentation (#540)
* Workflows: ignore if documentation files are changed

* Added nav info

* Fixed mkdocs.yml

* Removed regex in workflows

* Acknowledgments

* Added unification_framework.md

* Added support for MathJax

* Trying optimization problem description with LaTeX array

* Fixed optimization problems

* Added Unified view of state-of-the-art techniques

* Fixed missing \end{array} and replaced PDF image with PNG

* Added support for footnotes

* Added section Uno solver

* Trying new nav

* Added footer and link to Github

* Fixed syntax in mkdocs.yml

* Added 
[…]
