+++ docs/acknowledgments.md
@@ -1,0 +1,7 @@
+# Acknowledgments
+
+This work was supported by the Applied Mathematics activity within U.S. Department of Energy, Office of Science, Advanced Scientific Computing Research, under Contract DE-AC02-06CH11357.
+
+We would like to thank David Kiessling (KU Leuven) for fruitful discussions about constrained optimization and his extensive testing of Uno, Nils-Christian Kempke (Cardinal Operations) for his valuable help with technical questions in C++, and Gail Pieper (Argonne National Laboratory) for proofreading our manuscript.
+
+We are highly grateful to the following individuals who contributed their expertise and time to the open-source implementation of Uno: Oscar Dowson, Marcel Jacobse, David Kiessling, Rujia Liu, Stefano Lovato, Alexis Montoison, Manuel Schaich, and Silvio Traversaro. We have greatly benefited from their careful contributions and discussions.
