+++ .github/workflows/julia-tests-ubuntu.yml
@@ -6,12 +6,18 @@ on:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.yml'
+      - '*.yaml'
+      - 'docs/**'
   pull_request:
     branches: [main]
     paths-ignore:
       - '*.md'
       - 'LICENSE'
       - '*.cff'
+      - '*.yml'
+      - '*.yaml'
+      - 'docs/**'
     
 # needed to allow julia-actions/cache to delete old caches that it has created
 permissions:
