### 2026-03-02 3505ed0
Link SPRAL_jll's libspral in Yggdrasil build + run Julia tests with SSIDS in CI
