+++ .github/workflows/julia-tests-ubuntu.yml
@@ -29,7 +29,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        UNO_PRESET: [uno_filtersqp_bqpd, uno_filtersqp_highs, uno_filtersqp_bqpd_identity, uno_filtersqp_bqpd_regularization_tr, uno_filtersqp_bqpd_regularization_ls, uno_filtersqp_bqpd_ls_lbfgs, uno_ipopt_mumps, uno_ipopt_mumps_identity, uno_filterslp_bqpd, uno_filterslp_highs, uno_filtersqp_bqpd_charlies_instances]
+        UNO_PRESET: [uno_filtersqp_bqpd, uno_filtersqp_highs, uno_filtersqp_bqpd_identity, uno_filtersqp_bqpd_regularization_tr, uno_filtersqp_bqpd_regularization_ls, uno_filtersqp_bqpd_ls_lbfgs, uno_ipopt_mumps, uno_ipopt_mumps_identity, uno_ipopt_ssids, uno_filterslp_bqpd, uno_filterslp_highs, uno_filtersqp_bqpd_charlies_instances]
     name: Julia tests of ${{matrix.UNO_PRESET}} - ${{ github.event_name }}
     
     steps:
