### 2026-03-02 fbe88cb
Set logger=INFO in uno_ipopt_ssids.jl (most likely SSIDS is not found)
