### 2026-03-02 6d475ce
Set env variables for SPRAL in Yggdrasil build
