### 2026-03-03 a58f75d
Added workflow that installs unopy from PyPI and runs the example
