+++ .github/workflows/unopy-test-pypi.yml
@@ -27,11 +27,14 @@ jobs:
   build:
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest]
         python-version: ["3.12"]
 
     steps:
+      - uses: actions/checkout@v4
+    
       - name: Set up Python
         # This is the version of the action for setting up Python, not the Python version.
         uses: actions/setup-python@v5
