### 2026-03-03 f2368bf
Build macOS wheels with clang/clang++/gfortran instead of full GNU
