+++ .github/workflows/build-python-wheels.yml
@@ -50,7 +50,11 @@ jobs:
         run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
         env:
           MACOSX_DEPLOYMENT_TARGET: 15.0
-
+          CIBW_ENVIRONMENT_MACOS: >
+            CMAKE_C_COMPILER=clang
+            CMAKE_CXX_COMPILER=clang++
+            CMAKE_Fortran_COMPILER=gfortran-15
+            
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
       - uses: actions/upload-artifact@v4
