+++ .github/workflows/build-python-wheels.yml
@@ -44,7 +44,7 @@ jobs:
         uses: fortran-lang/setup-fortran@main
         with:
           compiler: 'gcc'
-          version: '14'
+          version: '15'
           
       - name: Install libomp (macOS)
         if: startsWith(matrix.os, 'macos')
@@ -79,12 +79,12 @@ jobs:
             CMAKE_C_COMPILER=clang
             CMAKE_CXX_COMPILER=clang++
             CMAKE_Fortran_COMPILER=gfortran-15
-            CMAKE_EXE_LINKER_FLAGS="-L${{env.LIBGFORTRAN_FOLDER}}"
-            CMAKE_SHARED_LINKER_FLAGS="-L${{env.LIBGFORTRAN_FOLDER}}"
-            OpenMP_C_FLAGS="-Xpreprocessor -fopenmp"
-            OpenMP_C_LIB_NAMES="omp"
-            OpenMP_CXX_FLAGS="-Xpreprocessor -fopenmp"
-            OpenMP_CXX_LIB_NAMES="omp"
+            CMAKE_EXE_LINKER_FLAGS=-L${{env.LIBGFORTRAN_FOLDER}}
+            CMAKE_SHARED_LINKER_FLAGS=-L${{env.LIBGFORTRAN_FOLDER}}
+            OpenMP_C_FLAGS=-Xpreprocessor\ -fopenmp
+            OpenMP_C_LIB_NAMES=omp
+            OpenMP_CXX_FLAGS=-Xpreprocessor\ -fopenmp
+            OpenMP_CXX_LIB_NAMES=omp
             OpenMP_omp_LIBRARY=${{ env.OPENMP_LIB }}
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
