### 2026-03-03 c53b6cc
Trying with Clang toolchain file
