### 2026-03-03 be93d97
[unopy] Release v0.2.2 (fixed macOS wheels) (#543)
* Build macOS wheels with clang/clang++/gfortran instead of full GNU

* Forces clang and clang++

* Install libomp

* Try without double quotes

* Trying with Clang toolchain file

* Fixes

* Moved linker flags to toolchain file

* [unopy] Test release v0.2.2
