+++ .github/workflows/build-python-wheels.yml
@@ -40,17 +40,56 @@ jobs:
       - run: pip install cibuildwheel
 
       - name: Install Fortran compiler on macOS
+        if: startsWith(matrix.os, 'macos')
         uses: fortran-lang/setup-fortran@main
         with:
           compiler: 'gcc'
-          version: '13'
+          version: '14'
+          
+      - name: Install libomp (macOS)
         if: startsWith(matrix.os, 'macos')
+        run: |
+          brew install libomp
+          if [[ $(uname -m) == 'arm64' ]]; then
+            echo "OPENMP_LIB=/opt/homebrew/opt/libomp/lib/libomp.dylib" >> $GITHUB_ENV
+            echo "LIBGFORTRAN_FOLDER=/opt/homebrew/opt/gcc/lib/gcc/15" >> $GITHUB_ENV
+          else
+            echo "OPENMP_LIB=/usr/local/opt/libomp/lib/libomp.dylib" >> $GITHUB_ENV
+            echo "LIBGFORTRAN_FOLDER=/usr/local/opt/gcc/lib/gcc/15" >> $GITHUB_ENV
+          fi
 
-      - name: Build wheels
+      - name: Build wheels (Linux)
+        if: runner.os == 'Linux'
         run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
         env:
-          MACOSX_DEPLOYMENT_TARGET: 15.0
+          CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
+      - name: Build wheels (macOS)
+        if: startsWith(matrix.os, 'macos')
+        run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
+        env:
+          # Override what setup-fortran set at the step level
+          CC: clang
+          CXX: clang++
+          MACOSX_DEPLOYMENT_TARGET: 15.0
+          CIBW_BEFORE_ALL_MACOS: |
+            cat > /tmp/clang_toolchain.cmake << 'EOF'
+            set(CMAKE_C_COMPILER clang)
+            set(CMAKE_CXX_COMPILER clang++)
+            set(CMAKE_Fortran_COMPILER gfortran-15)
+            set(OpenMP_C_FLAGS "-Xpreprocessor -fopenmp")
+            set(OpenMP_CXX_FLAGS "-Xpreprocessor -fopenmp")
+            set(OpenMP_C_LIB_NAMES omp)
+            set(OpenMP_CXX_LIB_NAMES omp)
+            set(OpenMP_omp_LIBRARY $OPENMP_LIB)
+            set(CMAKE_EXE_LINKER_FLAGS "-L$LIBGFORTRAN_FOLDER -lgfortran -Wl,-rpath,$LIBGFORTRAN_FOLDER")
+            set(CMAKE_SHARED_LINKER_FLAGS "-L$LIBGFORTRAN_FOLDER -lgfortran -Wl,-rpath,$LIBGFORTRAN_FOLDER")
+            EOF
+          CIBW_ENVIRONMENT_MACOS: >
+            CC=clang
+            CXX=clang++
+            FC=gfortran-15
+            CMAKE_TOOLCHAIN_FILE=/tmp/clang_toolchain.cmake
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
       - uses: actions/upload-artifact@v4
