+++ .github/workflows/unopy-test-pypi.yml
@@ -1,51 +1,0 @@
-name: PyPI's unopy on Python example
-
-on:
-  push:
-    branches: [ "main" ]
-    paths-ignore:
-      - '*.md'
-      - 'LICENSE'
-      - '*.cff'
-      - '*.yml'
-      - '*.yaml'
-      - 'docs/**'
-  pull_request:
-    branches: [ "main" ]
-    paths-ignore:
-      - '*.md'
-      - 'LICENSE'
-      - '*.cff'
-      - '*.yml'
-      - '*.yaml'
-      - 'docs/**'
-
-env:
-  BUILD_TYPE: Debug
-
-jobs:
-  build:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest, macos-latest]
-        python-version: ["3.12"]
-
-    steps:
-      - uses: actions/checkout@v4
-    
-      - name: Set up Python
-        # This is the version of the action for setting up Python, not the Python version.
-        uses: actions/setup-python@v5
-        with:
-          # Semantic version range syntax or exact version of a Python version
-          python-version: ${{ matrix.python-version }}
-          architecture: ${{ matrix.architecture }}
-
-      - name: Install unopy from PyPI
-        run: pip install unopy
-
-      - name: Run example
-        working-directory: ${{github.workspace}}/interfaces/Python/example
-        run: python example_hs015.py
