### 2026-03-03 2fd18d1
[CI] Test PyPI's unopy (#542)
* Added workflow that installs unopy from PyPI and runs the example

* Fixes

* Moved test of PyPI's artifacts after upload to PyPI
