### 2026-03-03 ab8cd5d
Simplify problem type assignment logic
Removed redundant check for hessian_sparsity when setting problem_type to 'LP'.
