### 2026-03-03 105b43f
Bypass dual_multiplier
