### 2026-03-03 822e239
UserModel.hpp: initialize members with {} instead of {nullptr}
