### 2026-03-03 e296062
PythonModel: fixed boolean functions by replacing != nullptr with .has_value()
