### 2026-03-03 d652a79
[unopy] Release v0.2.3 (fixed access to model components in `PythonModel`) (#547)
* UserModel.hpp: initialize members with {} instead of {nullptr}

* PythonModel: fixed boolean functions by replacing != nullptr with .has_value()

* [unopy] Test release v0.2.3
