+++ pyproject.toml
@@ -7,9 +7,9 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "unopy"
-version = "0.2.2"
+version = "0.2.3"
 #requires-python = ">=3.8"
-description = "Uno, a highly customizable SQP & barrier solver for nonlinearly constrained optimization"
+description = "With Uno, finally take full control of your SQP/barrier solver for nonlinearly constrained optimization"
 readme = "interfaces/Python/README.md"
 authors = [
     {name = "Charlie Vanaret", email = "vanaret@zib.de"},
