### 2026-03-03 e96291b
Replaced PointerWrapper and Vector with native py::array_t
