+++ interfaces/Python/cpp_classes/PythonModel.cpp
@@ -1,6 +1,9 @@
 // Copyright (c) 2025 Charlie Vanaret
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
+#include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
+#include <functional>
 #include "PythonModel.hpp"
 #include "optimization/EvaluationErrors.hpp"
 #include "symbolic/Concatenation.hpp"
@@ -44,9 +47,11 @@ namespace uno {
    double PythonModel::evaluate_objective(const Vector<double>& x) const {
       double objective_value = 0.;
       if (this->user_model.objective_function.has_value()) {
-         const py::object user_data = this->user_model.user_data.has_value() ? *this->user_model.user_data : py::cast(nullptr);
-         const uno_int return_code = (*this->user_model.objective_function)(static_cast<uno_int>(this->number_variables),
-            x, &objective_value, user_data);
+         const auto x_pybind = to_const_array(x.data(), this->number_variables);
+         auto objective_pybind = to_array(&objective_value, 1);
+
+         // evaluate objective
+         const uno_int return_code = (*this->user_model.objective_function)(x_pybind, objective_pybind);
          if (0 < return_code) {
             throw FunctionEvaluationError();
          }
@@ -58,9 +63,11 @@ namespace uno {
 
    void PythonModel::evaluate_constraints(const Vector<double>& x, Vector<double>& constraints) const {
       if (this->user_model.constraint_functions.has_value()) {
-         const py::object user_data = this->user_model.user_data.has_value() ? *this->user_model.user_data : py::cast(nullptr);
-         const uno_int return_code = (*this->user_model.constraint_functions)(static_cast<uno_int>(this->number_variables),
-            static_cast<uno_int>(this->number_constraints), x, constraints.data(), user_data);
+         const auto x_pybind = to_const_array(x.data(), this->number_variables);
+         auto constraints_pybind = to_array(constraints.data(), this->number_constraints);
+
+         // evaluate constraints
+         const uno_int return_code = (*this->user_model.constraint_functions)(x_pybind, constraints_pybind);
          if (0 < return_code) {
             throw FunctionEvaluationError();
          }
@@ -70,9 +77,11 @@ namespace uno {
 
    void PythonModel::evaluate_objective_gradient(const Vector<double>& x, Vector<double>& gradient) const {
       if (this->user_model.objective_gradient.has_value()) {
-         const py::object user_data = this->user_model.user_data.has_value() ? *this->user_model.user_data : py::cast(nullptr);
-         const uno_int return_code = (*this->user_model.objective_gradient)(static_cast<uno_int>(this->number_variables),
-            x, gradient.data(), user_data);
+         const auto x_pybind = to_const_array(x.data(), this->number_variables);
+         auto gradient_pybind = to_array(gradient.data(), this->number_variables);
+
+         // evaluate objective gradient
+         const uno_int return_code = (*this->user_model.objective_gradient)(x_pybind, gradient_pybind);
          if (0 < return_code) {
             throw GradientEvaluationError();
          }
@@ -84,7 +93,7 @@ namespace uno {
    }
 
    void PythonModel::compute_jacobian_sparsity(uno_int * row_indices, uno_int * column_indices, uno_int solver_indexing,
-                                                          MatrixOrder /*matrix_format*/) const {
+         MatrixOrder /*matrix_format*/) const {
       // copy the indices of the user sparsity patterns to the Uno vectors
       std::copy_n(this->user_model.jacobian_row_indices.data(), static_cast<size_t>(this->user_model.number_jacobian_nonzeros), row_indices);
[…]
