### 2026-03-03 9779c9b
Replaced PointerWrapper and Vector with native py::array_t
