+++ interfaces/Python/unopy.cpp
@@ -12,7 +12,6 @@ namespace py = pybind11;
 
 namespace uno {
    void define_Vector(py::module& module);
-   void define_PointerWrapper(py::module& module);
    void define_Model(py::module& module);
    void define_Result(py::module& module);
    void define_UnoSolver(py::module& module);
@@ -26,7 +25,6 @@ namespace uno {
       module.doc() = description;
 
       define_Vector(module);
-      define_PointerWrapper(module);
       define_Model(module);
       define_Result(module);
       define_UnoSolver(module);
