+++ interfaces/Python/unopy.hpp
@@ -8,32 +8,57 @@
 #include <optional>
 #include <vector>
 #include <pybind11/pybind11.h>
+#include <pybind11/numpy.h>
 #include "../UserModel.hpp"
-#include "linear_algebra/Vector.hpp"
-#include "tools/PointerWrapper.hpp"
 
 namespace py = pybind11;
 
 namespace uno {
-   using Objective = std::function<int(uno_int, const Vector<double>&, PointerWrapper<double>, const py::object&)>;
+   // alias for a typed, C-contiguous, non-owning numpy array view.
+   // Use for INPUT: zero-copy if the source array already matches.
+   template <typename T>
+   using const_array = py::array_t<T, py::array::c_style | py::array::forcecast>;
 
-   using Constraints = std::function<int(uno_int, uno_int, const Vector<double>&, PointerWrapper<double>, const py::object&)>;
+   // for OUTPUT buffers owned by C++: wrap the raw pointer in a numpy array
+   // with an explicit no-op capsule so Python knows not to free it.
+   template <typename T>
+   py::array_t<T> to_array(T* ptr, size_t n) {
+      return py::array_t<T>(
+          {static_cast<py::ssize_t>(n)}, // shape
+          {sizeof(T)}, // strides
+          ptr, // data (NOT owned)
+          py::capsule(ptr, [](void*) {}) // no-op deleter: C++ owns this memory
+      );
+   }
 
-   using ObjectiveGradient = std::function<int(uno_int, const Vector<double>&, PointerWrapper<double>, const py::object&)>;
+   template <typename T>
+   py::array_t<T> to_const_array(const T* ptr, size_t n) {
+      auto array = py::array_t<double>(
+          {static_cast<py::ssize_t>(n)},
+          {sizeof(T)},
+          const_cast<T*>(ptr),
+          py::capsule(ptr, [](void*) {})
+      );
+      array.attr("flags").attr("writeable") = false;
+      return array;
+   }
 
-   using Jacobian = std::function<int(uno_int, uno_int, const Vector<double>&, PointerWrapper<double>, const py::object&)>;
+   using Objective = std::function<int(const_array<double>, py::array_t<double>)>;
 
-   using Hessian = std::function<int(uno_int, uno_int, uno_int, const Vector<double>&, double, const Vector<double>&,
-      PointerWrapper<double>, const py::object&)>;
+   using Constraints = std::function<int(const_array<double>, py::array_t<double>)>;
 
-   using JacobianOperator = std::function<int(uno_int, uno_int, PointerWrapper<const double>, bool, PointerWrapper<const double>,
-      PointerWrapper<double>, const py::object&)>;
+   using ObjectiveGradient = std::function<int(const_array<double>, py::array_t<double>)>;
 
-   using JacobianTransposedOperator = std::function<int(uno_int, uno_int, PointerWrapper<const double>, bool,
-      PointerWrapper<const double>, PointerWrapper<double>, const py::object&)>;
+   using Jacobian = std::function<int(const_array<double>, py::array_t<double>)>;
 
-   using HessianOperator = std::function<int(uno_int, uno_int, PointerWrapper<const double>, bool, double,
-      const Vector<double>&, PointerWrapper<const double>, PointerWrapper<double>, const py::object&)>;
+   using Hessian = std::function<int(const_array<double>, double, const_array<double>, py::array_t<double>)>;
+
+   using JacobianOperator = std::function<int(const_array<double>, bool, const_array<double>, py::array_t<double>)>;
+
+   using JacobianTransposedOperator = std::function<int(const_array<double>, bool, const_array<double>, py::array_t<double>)>;
+
[…]
