### 2026-03-03 d30a3c1
Apply suggestion from @amontoison
