+++ interfaces/Python/cpp_classes/PythonModel.cpp
@@ -47,11 +47,11 @@ namespace uno {
    double PythonModel::evaluate_objective(const Vector<double>& x) const {
       double objective_value = 0.;
       if (this->user_model.objective_function.has_value()) {
-         const auto x_pybind = to_const_array(x.data(), this->number_variables);
-         auto objective_pybind = to_array(&objective_value, 1);
+         const auto x_py = to_const_array(x.data(), this->number_variables);
+         auto objective_py = to_array(&objective_value, 1);
 
          // evaluate objective
-         const uno_int return_code = (*this->user_model.objective_function)(x_pybind, objective_pybind);
+         const uno_int return_code = (*this->user_model.objective_function)(x_py, objective_py);
          if (0 < return_code) {
             throw FunctionEvaluationError();
          }
@@ -63,11 +63,11 @@ namespace uno {
 
    void PythonModel::evaluate_constraints(const Vector<double>& x, Vector<double>& constraints) const {
       if (this->user_model.constraint_functions.has_value()) {
-         const auto x_pybind = to_const_array(x.data(), this->number_variables);
-         auto constraints_pybind = to_array(constraints.data(), this->number_constraints);
+         const auto x_py = to_const_array(x.data(), this->number_variables);
+         auto constraints_py = to_array(constraints.data(), this->number_constraints);
 
          // evaluate constraints
-         const uno_int return_code = (*this->user_model.constraint_functions)(x_pybind, constraints_pybind);
+         const uno_int return_code = (*this->user_model.constraint_functions)(x_py, constraints_py);
          if (0 < return_code) {
             throw FunctionEvaluationError();
          }
@@ -77,11 +77,11 @@ namespace uno {
 
    void PythonModel::evaluate_objective_gradient(const Vector<double>& x, Vector<double>& gradient) const {
       if (this->user_model.objective_gradient.has_value()) {
-         const auto x_pybind = to_const_array(x.data(), this->number_variables);
-         auto gradient_pybind = to_array(gradient.data(), this->number_variables);
+         const auto x_py = to_const_array(x.data(), this->number_variables);
+         auto gradient_py = to_array(gradient.data(), this->number_variables);
 
          // evaluate objective gradient
-         const uno_int return_code = (*this->user_model.objective_gradient)(x_pybind, gradient_pybind);
+         const uno_int return_code = (*this->user_model.objective_gradient)(x_py, gradient_py);
          if (0 < return_code) {
             throw GradientEvaluationError();
          }
@@ -127,11 +127,11 @@ namespace uno {
 
    void PythonModel::evaluate_jacobian(const Vector<double>& x, double* jacobian_values) const {
       if (this->user_model.jacobian.has_value()) {
-         const auto x_pybind = to_const_array(x.data(), this->number_variables);
-         auto jacobian_pybind = to_array(jacobian_values, this->number_jacobian_nonzeros());
+         const auto x_py = to_const_array(x.data(), this->number_variables);
+         auto jacobian_py = to_array(jacobian_values, this->number_jacobian_nonzeros());
 
          // evaluate Jacobian
-         const uno_int return_code = (*this->user_model.jacobian)(x_pybind, jacobian_pybind);
+         const uno_int return_code = (*this->user_model.jacobian)(x_py, jacobian_py);
          if (0 < return_code) {
             throw GradientEvaluationError();
          }
@@ -147,13 +147,13 @@ namespace uno {
          if (this->user_model.lagrangian_sign_convention == UNO_MULTIPLIER_POSITIVE) {
             const_cast<Vector<double>&>(multipliers).scale(-1.);
          }
-         const auto x_pybind = to_const_array(x.data(), this->number_variables);
[…]
