### 2026-03-04 bc3d226
Model module: move vectors and callbacks into Model instead of copying
