### 2026-03-04 67e764f
Model module: move vectors and callbacks into Model instead of copying
