+++ interfaces/Python/python_modules/Model.cpp
@@ -5,68 +5,52 @@
 #include <pybind11/functional.h>
 #include <pybind11/stl.h>
 #include "../unopy.hpp"
-#include "linear_algebra/Vector.hpp"
-#include "symbolic/Range.hpp"
 
 namespace py = pybind11;
-
-PYBIND11_MAKE_OPAQUE(uno::Vector<double>)
 
 namespace uno {
    void define_Model(py::module& module) {
       py::class_<PythonUserModel>(module, "Model")
       // constructor
       // https://stackoverflow.com/a/62310838/16037994
-      .def(py::init<>([](const char* problem_type, uno_int number_variables, const std::vector<double>& variables_lower_bounds,
-            const std::vector<double>& variables_upper_bounds, uno_int base_indexing) {
+      .def(py::init<>([](const char* problem_type, uno_int number_variables, std::vector<double> variables_lower_bounds,
+            std::vector<double> variables_upper_bounds, uno_int base_indexing) {
          PythonUserModel model(problem_type, number_variables, base_indexing);
-         // copy the bounds internally
-         const size_t unsigned_number_variables = static_cast<size_t>(number_variables);
-         model.variables_lower_bounds = std::vector<double>(unsigned_number_variables);
-         model.variables_upper_bounds = std::vector<double>(unsigned_number_variables);
-         for (size_t variable_index: Range(unsigned_number_variables)) {
-            (*model.variables_lower_bounds)[variable_index] = variables_lower_bounds[variable_index];
-            (*model.variables_upper_bounds)[variable_index] = variables_upper_bounds[variable_index];
-         }
+         model.variables_lower_bounds = std::move(variables_lower_bounds);
+         model.variables_upper_bounds = std::move(variables_upper_bounds);
          return model;
       }), "Constructor")
 
       // methods
-      .def("set_objective", [](PythonUserModel& user_model, uno_int optimization_sense, const Objective& objective_function,
-            const ObjectiveGradient& objective_gradient) {
+      .def("set_objective", [](PythonUserModel& user_model, uno_int optimization_sense, Objective objective_function,
+            ObjectiveGradient objective_gradient) {
          if (optimization_sense != UNO_MINIMIZE && optimization_sense != UNO_MAXIMIZE) {
             throw std::runtime_error("Please specify a valid objective sense.");
          }
          user_model.optimization_sense = optimization_sense;
-         user_model.objective_function = objective_function;
-         user_model.objective_gradient = objective_gradient;
+         user_model.objective_function = std::move(objective_function);
+         user_model.objective_gradient = std::move(objective_gradient);
       })
 
-      .def("set_constraints", [](PythonUserModel& user_model, uno_int number_constraints, const Constraints& constraint_functions,
-            std::vector<double>& constraints_lower_bounds, std::vector<double>& constraints_upper_bounds, uno_int number_jacobian_nonzeros,
-            const std::vector<uno_int>& jacobian_row_indices, const std::vector<uno_int>& jacobian_column_indices,
-            const Jacobian& jacobian) {
+      .def("set_constraints", [](PythonUserModel& user_model, uno_int number_constraints, Constraints constraint_functions,
+            std::vector<double> constraints_lower_bounds, std::vector<double> constraints_upper_bounds, uno_int number_jacobian_nonzeros,
+            std::vector<uno_int> jacobian_row_indices, std::vector<uno_int> jacobian_column_indices, Jacobian jacobian) {
          if (number_constraints <= 0) {
             throw std::runtime_error("Please specify a positive number of constraints.");
          }
          user_model.number_constraints = number_constraints;
[…]
