+++ interfaces/Python/unopy.cpp
@@ -11,7 +11,6 @@ namespace py = pybind11;
 //PYBIND11_MAKE_OPAQUE(uno::Vector<double>)
 
 namespace uno {
-   void define_Vector(py::module& module);
    void define_Model(py::module& module);
    void define_Result(py::module& module);
    void define_UnoSolver(py::module& module);
@@ -24,7 +23,6 @@ namespace uno {
       description.append(", a solver for nonlinearly constrained optimization");
       module.doc() = description;
 
-      define_Vector(module);
       define_Model(module);
       define_Result(module);
       define_UnoSolver(module);
