### 2026-03-04 f7a9af4
Added some more checks + cleaned up a few things
