### 2026-03-04 63279c2
Added some more checks + cleaned up a few things
