+++ interfaces/Python/python_modules/UnoSolver.cpp
@@ -39,5 +39,5 @@ namespace uno {
       .def("optimize", [](UnoSolverWrapper& solver, const PythonUserModel& user_model) {
          return solver.optimize(user_model);
       }, py::arg("model"), "Optimize an optimization model with the Uno solver");
-}
+   }
 } // namespace
