+++ .github/workflows/build-python-wheels.yml
@@ -47,11 +47,7 @@ jobs:
           python-version: 3.11
 
       - run: pip install cibuildwheel
-
-      - name: Provide gfortran (Windows)
-        if: runner.os == 'Windows'
-        uses: msys2/setup-msys2@v2
-
+      
       - name: Install Fortran compiler (macOS)
         if: startsWith(matrix.os, 'macos')
         uses: fortran-lang/setup-fortran@main
