+++ .github/workflows/build-python-wheels.yml
@@ -49,14 +49,14 @@ jobs:
       - run: pip install cibuildwheel
       
       - name: Install Fortran compiler (macOS)
-        if: startsWith(matrix.os, 'macos')
+        if: startsWith(matrix.sys.os, 'macos')
         uses: fortran-lang/setup-fortran@main
         with:
           compiler: 'gcc'
           version: '14'
           
       - name: Install libomp (macOS)
-        if: startsWith(matrix.os, 'macos')
+        if: startsWith(matrix.sys.os, 'macos')
         run: |
           brew install libomp
           if [[ $(uname -m) == 'arm64' ]]; then
@@ -95,7 +95,7 @@ jobs:
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
       - name: Build wheels (macOS)
-        if: startsWith(matrix.os, 'macos')
+        if: startsWith(matrix.sys.os, 'macos')
         run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
         env:
           # Override what setup-fortran set at the step level
@@ -231,7 +231,7 @@ jobs:
   test_pypi:
     needs: upload_pypi
     name: Test PyPI artifacts
-    runs-on: ${{ matrix.os }}
+    runs-on: ${{ matrix.sys.os }}
     strategy:
       fail-fast: false
       matrix:
