+++ .github/workflows/build-python-wheels.yml
@@ -41,12 +41,6 @@ jobs:
 
     steps:
       - uses: actions/checkout@v5
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: 3.11
-
-      - run: pip install cibuildwheel
       
       - name: Install Fortran compiler (macOS)
         if: startsWith(matrix.sys.os, 'macos')
@@ -90,13 +84,13 @@ jobs:
 
       - name: Build wheels (Linux)
         if: runner.os == 'Linux'
-        run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
+        uses: pypa/cibuildwheel@v2.22.0
         env:
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
       - name: Build wheels (macOS)
         if: startsWith(matrix.sys.os, 'macos')
-        run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
+        uses: pypa/cibuildwheel@v2.22.0
         env:
           # Override what setup-fortran set at the step level
           CC: clang
@@ -124,7 +118,7 @@ jobs:
       
       - name: Build wheels (Windows)
         if: runner.os == 'Windows'
-        run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
+        uses: pypa/cibuildwheel@v2.22.0
         env:
           CIBW_BEFORE_BUILD: bash {package}/dependencies/scripts/download_dependencies.sh
           CIBW_BEFORE_BUILD_WINDOWS: >
