+++ .github/workflows/build-python-wheels.yml
@@ -63,12 +63,18 @@ jobs:
       - name: Build wheels (Linux)
         if: runner.os == 'Linux'
         uses: pypa/cibuildwheel@v2.22.0
+        with:
+          package-dir: .
+          output-dir: wheelhouse
         env:
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
 
       - name: Build wheels (macOS)
         if: startsWith(matrix.sys.os, 'macos')
         uses: pypa/cibuildwheel@v2.22.0
+        with:
+          package-dir: .
+          output-dir: wheelhouse
         env:
           # Override what setup-fortran set at the step level
           CC: clang
@@ -97,6 +103,9 @@ jobs:
       - name: Build wheels (Windows)
         if: runner.os == 'Windows'
         uses: pypa/cibuildwheel@v2.22.0
+        with:
+          package-dir: .
+          output-dir: wheelhouse
         env:
           CIBW_BEFORE_BUILD: bash {package}/dependencies/scripts/download_dependencies.sh
           CIBW_BEFORE_BUILD_WINDOWS: >
