### 2026-03-04 7e18276
Trying with CIBW_CONFIG_SETTINGS_WINDOWS
