+++ .github/workflows/build-python-wheels.yml
@@ -111,6 +111,9 @@ jobs:
             CMAKE_C_COMPILER=C:/mingw64/bin/gcc.exe
             CMAKE_CXX_COMPILER=C:/mingw64/bin/g++.exe
             CMAKE_Fortran_COMPILER=C:/mingw64/bin/gfortran.exe
+          CIBW_CONFIG_SETTINGS_WINDOWS: >
+            cmake.define.METIS_LIBRARY=dependencies/bin/libmetis.dll
+            cmake.define.HIGHS=dependencies/bin/libhighs.dll
           # Skip some builds
           CIBW_SKIP: cp36-* cp37-* cp*-*_i686 pp* *-win32
 
