### 2026-03-04 fdc6bdf
Do not fail fast
