### 2026-03-04 377f856
Compile SPRAL for the future releases
