+++ .github/julia/build_tarballs_release.jl
@@ -10,6 +10,10 @@ version = VersionNumber(ENV["UNO_RELEASE"])
 # Collection of sources required to complete build
 sources = [
     GitSource(ENV["UNO_URL"], ENV["UNO_COMMIT"]),
+    # v2025.9.18
+    GitSource("https://github.com/ralna/spral.git",
+              "e68988612dbd920323cee44c9cb8c6134847a990"),
+    # v5.8.2
     ArchiveSource("https://mumps-solver.org/MUMPS_5.8.2.tar.gz",
                   "eb515aa688e6dbab414bb6e889ff4c8b23f1691a843c68da5230a33ac4db7039"),
 ]
@@ -28,13 +32,19 @@ for file in $(ls .); do
       fi
    fi
 done
+
 cd ${prefix}
 cp -rL share/licenses deps/licenses
 mkdir deps/licenses/MUMPS
 cp $WORKSPACE/srcdir/MUMPS_5.8.2/LICENSE deps/licenses/MUMPS/LICENSE
+mkdir deps/licenses/spral
+cp $WORKSPACE/srcdir/spral/LICENCE deps/licenses/spral/LICENCE
 chmod -R u=rwx deps
 tar -czvf deps.tar.gz deps
 rm -r deps
+
+# Update Ninja
+cp ${host_prefix}/bin/ninja /usr/bin/ninja
 
 # Compile MUMPS
 mkdir -p ${libdir}
@@ -82,6 +92,24 @@ cp include/*.h ${includedir}
 cp libseq/*.h ${includedir}/libseq
 cp lib/*.${dlext} ${libdir}
 
+# Compile SPRAL
+cd $WORKSPACE/srcdir/spral
+
+if [[ "${target}" == *mingw* ]]; then
+  HWLOC="hwloc-15"
+else
+  HWLOC="hwloc"
+fi
+
+meson setup builddir --cross-file="${MESON_TARGET_TOOLCHAIN}" \
+                     --prefix=$prefix \
+                     -Dlibhwloc=$HWLOC \
+                     -Dlibblas=openblas \
+                     -Dliblapack=openblas
+
+meson compile -C builddir
+meson install -C builddir
+
 # Compile Uno
 cd $WORKSPACE/srcdir/Uno
 mkdir -p build
@@ -102,6 +130,7 @@ cmake \
     -DHIGHS=${libdir}/libhighs.${dlext} \
     -DBQPD=${prefix}/lib/libbqpd.a \
     -DHSL=${libdir}/libhsl.${dlext} \
+    -DSPRAL=${libdir}/libspral.${dlext} \
     -DMUMPS_INCLUDE_DIR=${includedir} \
     -DMETIS_INCLUDE_DIR=${includedir} \
     -DMUMPS_LIBRARY="${libdir}/libdmumps.${dlext}" \
@@ -128,6 +157,7 @@ platforms = expand_gfortran_versions(platforms)
 
 # The products that we will ensure are always built
 products = [
+   LibraryProduct("libspral", :libspral),
    LibraryProduct("libdmumps", :libdmumps),
    ExecutableProduct("uno_ampl", :amplexe),
    LibraryProduct("libuno", :libuno),
@@ -136,12 +166,14 @@ products = [
 
 # Dependencies that must be installed before this package can be built
 dependencies = [
+    HostBuildDependency(PackageSpec(name="Ninja_jll", uuid="76642167-d241-5cee-8c94-7a494e8cb7b7")),
     BuildDependency(PackageSpec(name="BQPD_jll", uuid="1325ac01-0a49-589f-8355-43321054aaab")),
[…]
