### 2026-03-04 ba915ce
Don't compile tests and examples of SPRAL
