### 2026-03-04 c1a0137
[unopy] Test release v0.2.4+1 (call cibuildwheel directly + test TestPyPI artifacts)
