+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
-    "scikit-build-core",
-    "pybind11"
+    "scikit-build-core>=0.8",
+    "pybind11>=2.11"
 ]
 build-backend = "scikit_build_core.build"
 
