+++ pyproject.toml
@@ -32,6 +32,8 @@ wheel.cmake = true
 
 # support for passing paths to dependencies from environment to CMake
 [tool.scikit-build.cmake.define]
+BLAS_LIBRARIES = "dependencies/lib/libblas.a;dependencies/lib/libcblas.a"
+LAPACK_LIBRARIES = "dependencies/lib/liblapack.a"
 # BQPD (cmake -DBQPD=/path/.../libbqpd.a)
 BQPD = "dependencies/lib/libbqpd.a"
 # MUMPS
@@ -42,7 +44,5 @@ MUMPS_LIBRARY = "dependencies/lib/libdmumps.a"
 MUMPS_COMMON_LIBRARY = "dependencies/lib/libmumps_common.a"
 MUMPS_PORD_LIBRARY = "dependencies/lib/libpord.a"
 MUMPS_MPISEQ_LIBRARY = "dependencies/lib/libmpiseq.a"
-BLAS_LIBRARIES = "dependencies/lib/libblas.a;dependencies/lib/libcblas.a"
-LAPACK_LIBRARIES = "dependencies/lib/liblapack.a"
 # HiGHS
 HIGHS = "dependencies/lib/libhighs.a"
