+++ .github/workflows/build-python-wheels.yml
@@ -23,31 +23,44 @@ on:
 
 jobs:
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
+    name: Build wheels on ${{ matrix.sys.os }}
+    runs-on: ${{ matrix.sys.os }}
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, ubuntu-22.04-arm, macos-latest, macos-15-intel]
+        sys:
+          - { os: ubuntu-latest, shell: bash }
+          - { os: ubuntu-22.04-arm, shell: bash }
+          - { os: macos-latest, shell: bash }
+          - { os: macos-15-intel, shell: bash }
+          - { os: windows-latest, shell: 'msys2 {0}' }
+    defaults:
+      run:
+        shell: ${{ matrix.sys.shell }}
 
     steps:
       - uses: actions/checkout@v5
-      
+
       - uses: actions/setup-python@v5
         with:
           python-version: "3.11"
 
       - run: pip install cibuildwheel
+        if: runner.os != 'Windows'
 
-      - name: Install Fortran compiler on macOS
-        if: startsWith(matrix.os, 'macos')
+      - run: pip install cibuildwheel
+        if: runner.os == 'Windows'
+        shell: pwsh
+
+      - name: Install Fortran compiler (macOS)
+        if: startsWith(matrix.sys.os, 'macos')
         uses: fortran-lang/setup-fortran@main
         with:
           compiler: 'gcc'
-          version: '14'
+          version: '13'
           
       - name: Install libomp (macOS)
-        if: startsWith(matrix.os, 'macos')
+        if: startsWith(matrix.sys.os, 'macos')
         run: |
           brew install libomp
           if [[ $(uname -m) == 'arm64' ]]; then
@@ -63,9 +76,11 @@ jobs:
         run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
         env:
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
+          # Skip some builds
+          CIBW_SKIP: cp36-* cp37-* cp*-*_i686 pp* *-win32
 
       - name: Build wheels (macOS)
-        if: startsWith(matrix.os, 'macos')
+        if: startsWith(matrix.sys.os, 'macos')
         run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
         env:
           # Override what setup-fortran set at the step level
@@ -91,11 +106,34 @@ jobs:
             FC=gfortran-15
             CMAKE_TOOLCHAIN_FILE=/tmp/clang_toolchain.cmake
           CIBW_BEFORE_BUILD: bash dependencies/scripts/download_dependencies.sh
+          # Skip some builds
+          CIBW_SKIP: cp36-* cp37-* cp*-*_i686 pp* *-win32
+      
+      - name: Build wheels (Windows)
+        if: runner.os == 'Windows'
+        run: cibuildwheel --output-dir ${{github.workspace}}/wheelhouse
+        shell: pwsh
+        env:
+          CIBW_BEFORE_BUILD: bash {package}/dependencies/scripts/download_dependencies.sh
+          CIBW_BEFORE_BUILD_WINDOWS: >
+            C:/msys64/usr/bin/bash -lc "cd \"$(cygpath -u '{package}')\" && ./dependencies/scripts/download_dependencies.sh"
+          CIBW_ENVIRONMENT_WINDOWS: >
+            CMAKE_GENERATOR="MinGW Makefiles"
+            CMAKE_MAKE_PROGRAM=C:/mingw64/bin/mingw32-make.exe
[…]
