+++ .github/workflows/unopy-testpypi-example.yml
@@ -40,22 +40,11 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Create virtual environment
-        run: python -m venv .venv
-
-      - name: Active venv (Windows)
-        if: runner.os == 'Windows'
-        run: .venv\Scripts\activate
-        
-      - name: Active venv
-        if: runner.os != 'Windows'
-        run: source .venv/bin/activate
-
       - name: Upgrade pip, setuptools, wheel
-        run: pip install --upgrade pip setuptools wheel
+        run: python -m pip install --upgrade pip setuptools wheel
       
       - name: Install unopy from TestPyPI
-        run: pip install -v --index-url https://test.pypi.org/simple/ --extra-index-url https://pypi.org/simple/ unopy
+        run: python -m pip install -v --index-url https://test.pypi.org/simple/ --extra-index-url https://pypi.org/simple/ unopy
 
       - name: Run example
         working-directory: ${{github.workspace}}/interfaces/Python/example
