### 2026-03-05 febcfa6
Add MSBuild/MSVC to PATH (Windows)
