### 2026-03-05 f066906
Diagnose unopy import (Windows)
