### 2026-03-05 298edb1
Inspect unopy on disk (Windows)
