+++ pyproject.toml
@@ -27,6 +27,7 @@ Issues = "https://github.com/cvanaret/Uno/issues"
 cmake.source-dir = "."
 cmake.verbose = true
 wheel.cmake = true
+wheel.install-dir = "." # installs .pyd directly into site-packages root
 
 # support for passing paths to dependencies from environment to CMake
 [tool.scikit-build.cmake.define]
