+++ interfaces/Python/unopy/__init__.py
@@ -1,0 +1,1 @@
+from .unopy import *  # re-export the .pyd contents
