+++ .github/workflows/build-python-wheels.yml
@@ -123,7 +123,7 @@ jobs:
           CIBW_BEFORE_BUILD_WINDOWS: >
             C:/msys64/usr/bin/bash -lc "cd \"$(cygpath -u '{package}')\" && ./dependencies/scripts/download_dependencies.sh"
           CIBW_REPAIR_WHEEL_COMMAND_WINDOWS: >
-            delvewheel repair --add-path dependencies/bin;dependencies/lib -w {dest_dir} {wheel}
+            delvewheel repair --add-path dependencies/bin;dependencies/lib --no-mangle-all -w {dest_dir} {wheel}
           CIBW_ENVIRONMENT_WINDOWS: >
             CMAKE_GENERATOR="MinGW Makefiles"
             CMAKE_MAKE_PROGRAM=C:/mingw64/bin/mingw32-make.exe
