### 2026-03-05 a12fd12
[unopy] Test release v0.2.8 (fixed path in pwsh)
