+++ .github/workflows/build-python-wheels.yml
@@ -125,7 +125,7 @@ jobs:
             CMAKE_Fortran_COMPILER=C:/mingw64/bin/gfortran.exe
           CIBW_CONFIG_SETTINGS_WINDOWS: >
             cmake.define.METIS_LIBRARY=dependencies/bin/libmetis.dll
-          # cmake.define.HIGHS=dependencies/bin/libhighs.dll
+            cmake.define.HIGHS=dependencies/bin/libhighs.dll
           # Skip some builds
           CIBW_SKIP: cp36-* cp37-* cp*-*_i686 pp* *-win32
 
