### 2026-03-05 977d1d8
Diagnose DLL issue (Windows)
