### 2026-03-05 c36a97c
Find missing symbol (Windows)
