### 2026-03-05 bf2b7ec
Find missing symbol (Windows)
