### 2026-03-05 c00c853
Find missing symbol (Windows)
