+++ .github/workflows/unopy-testpypi-example.yml
@@ -101,8 +101,8 @@ jobs:
             dll_name = entry.dll.decode()
             print('DLL: ' + dll_name)
               for imp in entry.imports:
-                  name = imp.name.decode() if imp.name else 'ordinal:' + str(imp.ordinal)
-                          print('  ' + name)
+                name = imp.name.decode() if imp.name else 'ordinal:' + str(imp.ordinal)
+                  print('  ' + name)
           '@
           $script | Out-File -FilePath diagnose.py -Encoding utf8
           python diagnose.py
