+++ CMakeLists.txt
@@ -322,7 +322,10 @@ if(SKBUILD)
    target_link_libraries(unopy PUBLIC ${LIBRARIES} ${FORTRAN_LIBS})
    target_include_directories(unopy SYSTEM PUBLIC ${DIRECTORIES})
 
-   install(TARGETS unopy LIBRARY DESTINATION ${CMAKE_INSTALL_PREFIX})
+   install(TARGETS unopy
+           LIBRARY DESTINATION unopy    # Linux/macOS: unopy/unopy.so
+           RUNTIME DESTINATION unopy    # Windows: unopy/unopy.pyd
+   )
 else()
    message(STATUS "SKBUILD not detected, skipping Python extension.")
 endif()
