### 2026-03-06 3760f47
Check DLL versions (Windows)
