### 2026-03-06 8e2780d
Check GCC versions in DLLs (Windows)
