### 2026-03-06 413c52b
Cross-reference imports vs exports (Windows)
