### 2026-03-06 5449ce1
Check system DLL imports (Windows)
