+++ CMakeLists.txt
@@ -243,6 +243,8 @@ if(APPLE)
    list(REMOVE_ITEM FORTRAN_LIBS gcc)
    list(REMOVE_ITEM FORTRAN_LIBS emutls_w)
    list(REMOVE_ITEM FORTRAN_LIBS heapt_w)
+elseif(WIN32)
+   list(REMOVE_ITEM FORTRAN_LIBS emutls_w)
 endif()
 
 # compile the Uno source files once, whatever the compilation type is (static or shared)
