### 2026-03-06 6653260
Added debug info in CMake
