### 2026-03-06 e358afa
Removed gcc from FORTRAN_LIBS for Windows
