### 2026-03-06 5978044
Check libgcc_s_seh exports (Windows)
