### 2026-03-06 0cabe5b
Find all libgcc_s_seh on PATH (Windows)
