### 2026-03-06 3188eaa
Prepend to PATH so our DLLs are found before any system MinGW ones
