### 2026-03-06 72fc913
Verify PATH and runtime DLL loading (Windows)
