### 2026-03-06 4339f8e
Diagnose libhighs.dll (Windows)
