### 2026-03-06 b3f0398
Exclude HiGHS
