+++ pyproject.toml
@@ -45,4 +45,4 @@ MUMPS_COMMON_LIBRARY = "dependencies/lib/libmumps_common.a"
 MUMPS_PORD_LIBRARY = "dependencies/lib/libpord.a"
 MUMPS_MPISEQ_LIBRARY = "dependencies/lib/libmpiseq.a"
 # HiGHS
-HIGHS = "dependencies/lib/libhighs.a"
+# HIGHS = "dependencies/lib/libhighs.a"
