### 2026-03-06 901d9d8
Temporarily removed HiGHS from example_hs015.py
