### 2026-03-06 dda1508
[unopy] build wheels for Windows (#480)
* Try building wheels for windows-11-arm

* Removed windows-11-arm, added windows-2022

* Install pkgconfiglite for Windows

* New attempt

* Install msys2/setup-msys2 for Windows

* Set environment variables

* Trying with rlalik/setup-cpp-compiler@master

* Try pdfo's workflow

* Use pypa/cibuildwheel@v2.0.0

* Remove Python install

* Relaxed constraint on scikit-build-core version

* Removed all constraints in pyproject.toml

* Reverting

* Try with delvewheel

* New attempt

* Added parameters to msys2/setup-msys2@v2

* Trying with pypa/cibuildwheel
[…]
