### 2026-03-06 a6fff11
[unopy] Release v0.2.4 (wheels for Windows)
