### 2026-03-06 9c493e9
No numpy install required in unopy-test-example.yml
