### 2026-03-08 3d9d09e
[unopy] Try to add HiGHS
