### 2026-03-08 54b49bb
[CI] Update build.yml to use static libhighs.a
