+++ .github/workflows/build.yml
@@ -161,6 +161,32 @@ jobs:
         nmake install
 
     # =========================================================
+    # MinGW: BUILD HiGHS
+    # =========================================================
+
+    - name: Build HiGHS (MinGW)
+      if: matrix.toolchain == 'mingw'
+      shell: cmd
+      run: |
+        git clone --branch v1.12.0 --depth 1 https://github.com/ERGO-Code/HiGHS.git
+        cd HiGHS
+        cmake -G "MinGW Makefiles" ^
+          -B build ^
+          -DCMAKE_INSTALL_PREFIX=${{github.workspace}}\dependencies ^
+          -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}} ^
+          -DBUILD_SHARED_LIBS=OFF ^
+          -DZLIB=OFF ^
+          -DHIPO=ON ^
+          -DBUILD_EXAMPLES=OFF ^
+          -DBUILD_TESTING=OFF ^
+          -DBUILD_CXX_EXE=OFF ^
+          -DMETIS_ROOT=${{github.workspace}}\dependencies ^
+          -DBLAS_LIBRARIES=${{github.workspace}}\dependencies\lib\libblas.a ^
+          -DCMAKE_POSITION_INDEPENDENT_CODE=ON ^
+          -DCMAKE_CXX_FLAGS="-static-libgcc -static-libstdc++"
+        cmake --build build --config ${{env.BUILD_TYPE}} -j4
+
+    # =========================================================
     # CONFIGURE
     # =========================================================
 
