### 2026-03-08 157a023
HiGHS doesn't like backslash for HIPO...
