+++ .github/workflows/build.yml
@@ -161,30 +161,41 @@ jobs:
         nmake install
 
     # =========================================================
+    # MinGW: Add C++ runtime
+    # =========================================================
+
+    - name: Download libstdc++.a
+      if: matrix.toolchain == 'mingw'
+      shell: bash
+      run: |
+        wget -O "${GITHUB_WORKSPACE}/tmp_libstdc++.tar.gz" "https://github.com/JuliaLang/PackageCompiler.jl/releases/download/v1.0.0/x86_64-8.1.0-release-posix-seh-rt_v6-rev0.tar.gz"
+        tar -xzf "${GITHUB_WORKSPACE}/tmp_libstdc++.tar.gz" -C "${GITHUB_WORKSPACE}/tmp_libstdc++"
+        cp "${GITHUB_WORKSPACE}/tmp_libstdc++/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libstdc++.a" "${GITHUB_WORKSPACE}/dependencies/lib/libstdc++.a"
+
+    # =========================================================
     # MinGW: BUILD HiGHS
     # =========================================================
 
-    - name: Build HiGHS (MinGW)
-      if: matrix.toolchain == 'mingw'
-      shell: cmd
-      run: |
-        git clone --branch v1.12.0 --depth 1 https://github.com/ERGO-Code/HiGHS.git
-        cd HiGHS
-        cmake -G "MinGW Makefiles" ^
-          -B build ^
-          -DCMAKE_INSTALL_PREFIX=${{github.workspace}}/dependencies ^
-          -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}} ^
-          -DBUILD_SHARED_LIBS=OFF ^
-          -DZLIB=OFF ^
-          -DHIPO=ON ^
-          -DBUILD_EXAMPLES=OFF ^
-          -DBUILD_TESTING=OFF ^
-          -DBUILD_CXX_EXE=OFF ^
-          -DMETIS_ROOT=${{github.workspace}}/dependencies ^
-          -DBLAS_LIBRARIES="D:/a/Uno/Uno/dependencies/lib/libblas.a" ^
-          -DCMAKE_POSITION_INDEPENDENT_CODE=ON ^
-          -DCMAKE_CXX_FLAGS="-static-libgcc -static-libstdc++"
-        cmake --build build --config ${{env.BUILD_TYPE}} -j4
+    # - name: Build HiGHS (MinGW)
+    #   if: matrix.toolchain == 'mingw'
+    #   shell: cmd
+    #   run: |
+    #     git clone --branch v1.12.0 --depth 1 https://github.com/ERGO-Code/HiGHS.git
+    #     cd HiGHS
+    #     cmake -G "MinGW Makefiles" ^
+    #       -B build ^
+    #       -DCMAKE_INSTALL_PREFIX=${{github.workspace}}/dependencies ^
+    #       -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}} ^
+    #       -DBUILD_SHARED_LIBS=OFF ^
+    #       -DZLIB=OFF ^
+    #       -DHIPO=ON ^
+    #       -DBUILD_EXAMPLES=OFF ^
+    #       -DBUILD_TESTING=OFF ^
+    #       -DBUILD_CXX_EXE=OFF ^
+    #       -DMETIS_ROOT=${{github.workspace}}/dependencies ^
+    #       -DBLAS_LIBRARIES="D:/a/Uno/Uno/dependencies/lib/libblas.a" ^
+    #       -DCMAKE_POSITION_INDEPENDENT_CODE=ON
+    #     cmake --build build --config ${{env.BUILD_TYPE}} -j4
 
[…]
