### 2026-03-08 3c53b3a
[CI] Update build.yml to use static libhighs.a (#554)
