### 2026-03-08 f1b293b
Add tests with MUMPS, SSIDS, MA27 and MA57 in UnoSolver.jl
