+++ interfaces/Julia/test/runtests.jl
@@ -1,6 +1,7 @@
 using UnoSolver
 using NLPModels, MathOptInterface, CUTEst
 using LinearAlgebra, Test
+using HSL_jll
 
 version = uno_version()
 println("The version of Uno is $version.")
