+++ interfaces/Julia/Project.toml
@@ -27,7 +27,8 @@ julia = "1.10"
 
 [extras]
 CUTEst = "1b53aba6-35b6-5f92-a507-53c67d53f819"
+HSL_jll = "017b0a0e-03f4-516a-9b91-836bbd1904dd"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["CUTEst", "MathOptInterface", "NLPModels", "Test"]
+test = ["CUTEst", "HSL_jll", "MathOptInterface", "NLPModels", "Test"]
