### 2026-03-08 b746235
[unopy] Test release v0.2.22 with HiGHS in the example file
