### 2026-03-08 aaa7b46
[unopy] Compile with HiGHS (#553)
* [unopy] Try to add HiGHS

* wip

* Update build_tarballs_utils.jl

* should be fine now!

* [unopy] Test release with HiGHS

* [unopy] Test release v0.2.22 with HiGHS in the example file

---------

Co-authored-by: Charlie Vanaret <vanaret@zib.de>
