+++ .github/julia/build_tarballs_utils.jl
@@ -1,4 +1,4 @@
-# julia +1.7 --color=yes build_tarballs_utils.jl x86_64-linux-gnu-libgfortran5,x86_64-apple-darwin-libgfortran5,x86_64-w64-mingw32-libgfortran5,aarch64-linux-gnu-libgfortran5,aarch64-apple-darwin-libgfortran5 --verbose --deploy="amontoison/UnoUtils_jll.jl"
+# julia +1.7 --color=yes build_tarballs_utils.jl x86_64-linux-gnu-libgfortran5-cxx11,x86_64-apple-darwin-libgfortran5-cxx11,x86_64-w64-mingw32-libgfortran5-cxx11,aarch64-linux-gnu-libgfortran5-cxx11,aarch64-apple-darwin-libgfortran5-cxx11 --verbose --deploy="amontoison/UnoUtils_jll.jl"
 using BinaryBuilder, Pkg
 
 name = "UnoUtils"
@@ -31,10 +31,10 @@ sources = [
     # HiGHS v1.12.0
     GitSource("https://github.com/ERGO-Code/HiGHS.git",
               "755a8e027a99a8d4ecf153a8dde4b2a767cdf384"),
-    # HiGHS v1.13.0
+    # HiGHS v1.13.1
     # GitSource("https://github.com/ERGO-Code/HiGHS.git",
-    #           "1bce6d5c801398dab6d2e6f98ac8935f3d4eec9c"),
-
+    #           "1d267d97c16928bb5f86fcb2cba2d20f94c8720c"),
+    #
     # Package compiler for Windows
     ArchiveSource("https://github.com/JuliaLang/PackageCompiler.jl/releases/download/v1.0.0/x86_64-8.1.0-release-posix-seh-rt_v6-rev0.tar.gz",
                   "fe3f401bc936fbe6af940b26c5e0f266f762a3416f979c706e599b24082dc5c7"),
@@ -174,7 +174,8 @@ cmake .. \
     -DBUILD_TESTING=OFF \
     -DBUILD_CXX_EXE=OFF \
     -DMETIS_ROOT=${prefix} \
-    -DBLAS_LIBRARIES=${prefix}/lib/libblas.a
+    -DBLAS_LIBRARIES=${prefix}/lib/libblas.a \
+    -DCMAKE_POSITION_INDEPENDENT_CODE=ON
 
 if [[ "${target}" == *-linux-* ]]; then
         make -j ${nproc}
@@ -186,6 +187,15 @@ else
     fi
 fi
 make install
+
+if [[ "${target}" == *-mingw* ]]; then
+    cp $WORKSPACE/srcdir/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libstdc++.a ${prefix}/lib/libstdc++.a
+    cp $WORKSPACE/srcdir/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libgfortran.a ${prefix}/lib/libgfortran.a
+    cp $WORKSPACE/srcdir/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libquadmath.a ${prefix}/lib/libquadmath.a
+    cp $WORKSPACE/srcdir/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libstdc++.a ${prefix}/lib/libgomp.a
+    cp $WORKSPACE/srcdir/mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libstdc++.a ${prefix}/lib/libgcc.a
[…]
