### 2026-03-08 3ce7bfe
[unopy] Release v0.2.5 (#557)
