+++ pyproject.toml
@@ -7,7 +7,10 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "unopy"
-version = "0.2.22"
+# PyPI
+version = "0.2.5"
+# TestPyPI
+#version = "0.2.22"
 description = "With Uno, finally take full control of your SQP/barrier solver for nonlinearly constrained optimization"
 readme = "interfaces/Python/README.md"
 authors = [
