### 2026-03-08 e7326f0
Print linear solver in NLPModels tests
