### 2026-03-08 e97d8f6
Correctly set the dimensions in the `Result` object (#558)
Correctly set the number of variables and constraints in the Result (not those of the reformulated model!)
