### 2026-03-08 669ece5
[UnoSolver.jl] Improve uno_get_solver_string_option
