+++ interfaces/Julia/test/C_wrapper.jl
@@ -227,6 +227,7 @@ y0 = zeros(Float64, ncon)
   solution_primal_feasibility = UnoSolver.uno_get_solution_primal_feasibility(solver)
   solution_stationarity = UnoSolver.uno_get_solution_stationarity(solver)
   solution_complementarity = UnoSolver.uno_get_solution_complementarity(solver)
+  hessian_model = UnoSolver.uno_get_solver_string_option(solver, "hessian_model")
 
   primal_solution = Vector{Float64}(undef, nvar)
   UnoSolver.uno_get_primal_solution(solver, primal_solution)
@@ -250,6 +251,75 @@ y0 = zeros(Float64, ncon)
   @test solution_primal_feasibility ≈ 0.0 atol = 1e-5
   @test solution_stationarity ≈ 0.0 atol = 1e-5
   @test solution_complementarity ≈ 0.0 atol = 1e-5
+  @test hessian_model == "exact"
+end
+
+@testset "L-BFGS with preset=filtersqp" begin
+  model = uno_model(
+    "NLP",
+    true,
+    nvar,
+    ncon,
+    lvar,
+    uvar,
+    lcon,
+    ucon,
+    jrows,
+    jcols,
+    nnzj,
+    hrows,
+    hcols,
+    nnzh,
+    c_objective,
+    c_constraints,
+    c_objective_gradient,
+    c_jacobian,
+    nothing,
+    c_jacobian_operator,
+    c_jacobian_transposed_operator,
+    nothing,
+  )
+
+  UnoSolver.uno_set_initial_primal_iterate(model, x0)
+  UnoSolver.uno_set_initial_dual_iterate(model, y0)
+
+  solver = uno_solver()
+  uno_set_solver_preset(solver, "filtersqp")
+  uno_set_solver_bool_option(solver, "print_solution", true)
+  uno_set_solver_string_option(solver, "hessian_model", "LBFGS")
+  uno_optimize(solver, model)
+
+  optimization_status = UnoSolver.uno_get_optimization_status(solver)
+  solution_status = UnoSolver.uno_get_solution_status(solver)
+  solution_objective = UnoSolver.uno_get_solution_objective(solver)
+  solution_primal_feasibility = UnoSolver.uno_get_solution_primal_feasibility(solver)
+  solution_stationarity = UnoSolver.uno_get_solution_stationarity(solver)
+  solution_complementarity = UnoSolver.uno_get_solution_complementarity(solver)
+  hessian_model = UnoSolver.uno_get_solver_string_option(solver, "hessian_model")
+
+  primal_solution = Vector{Float64}(undef, nvar)
+  UnoSolver.uno_get_primal_solution(solver, primal_solution)
+
+  constraint_dual_solution = Vector{Float64}(undef, ncon)
+  UnoSolver.uno_get_constraint_dual_solution(solver, constraint_dual_solution)
+
+  lower_bound_dual_solution = Vector{Float64}(undef, nvar)
+  UnoSolver.uno_get_lower_bound_dual_solution(solver, lower_bound_dual_solution)
+
+  upper_bound_dual_solution = Vector{Float64}(undef, nvar)
+  UnoSolver.uno_get_upper_bound_dual_solution(solver, upper_bound_dual_solution)
+
+  @test optimization_status == 0  # UNO_SUCCESS
+  @test solution_status == 1      # UNO_FEASIBLE_KKT_POINT
+  @test primal_solution[1] ≈ 1.0000000000000000 atol = 1e-5
+  @test primal_solution[2] ≈ 4.7429996418092970 atol = 1e-5
[…]
