### 2026-03-08 6870c10
[UnoSolver.jl] Support LBFGS with MathOptInterface / JuMP (#545)
