### 2026-03-08 1d48d97
[UnoSolver.jl] Release 0.2.4
