### 2026-03-08 112b17e
[UnoSolver.jl] Improve the support for unconstrained problems
