+++ uno/ingredients/globalization_mechanisms/GlobalizationMechanismFactory.hpp
@@ -15,7 +15,7 @@ namespace uno {
 
    class GlobalizationMechanismFactory {
    public:
-      static std::unique_ptr<GlobalizationMechanism> create(const Model& model, const Options& options);
+      static std::unique_ptr<GlobalizationMechanism> create(const Model& model, Options& options);
 
       constexpr static std::array available_strategies{"TR", "LS"};
    };
