+++ uno/ingredients/globalization_mechanisms/GlobalizationMechanism.cpp
@@ -13,7 +13,7 @@
 #include "tools/Statistics.hpp"
 
 namespace uno {
-   GlobalizationMechanism::GlobalizationMechanism(const Model& model, bool use_trust_region, const Options& options):
+   GlobalizationMechanism::GlobalizationMechanism(const Model& model, bool use_trust_region, Options& options):
       constraint_relaxation_strategy(ConstraintRelaxationStrategyFactory::create(model, use_trust_region, options)) {
    }
 
