### 2026-03-09 df45fe6
Simplified MOI_wrapper._quasi_newton_approximation
