### 2026-03-09 5f12956
Modify `Options` object when user defined options are overridden by Uno (#563)
* Modify Options file when user defined options are overridden by Uno

* Simplified MOI_wrapper._quasi_newton_approximation
