+++ uno/ingredients/constraint_relaxation_strategies/ConstraintRelaxationStrategyFactory.cpp
@@ -10,10 +10,12 @@
 
 namespace uno {
    std::unique_ptr<ConstraintRelaxationStrategy> ConstraintRelaxationStrategyFactory::create(const Model& model,
-         bool use_trust_region, const Options& options) {
+         bool use_trust_region, Options& options) {
       // figure out whether we need to relax constraints altogether
       if (model.number_constraints == 0) {
          INFO << "The model is unconstrained, picking no relaxation\n";
+         // override user defined option
+         options.set_string("constraint_relaxation_strategy", "no_relaxation", true);
          return std::make_unique<NoRelaxation>(model, options);
       }
       const std::string constraint_relaxation_type = options.get_string("constraint_relaxation_strategy");
