+++ docs/options.md
@@ -19,9 +19,9 @@ Defaults are taken from `uno/options/DefaultOptions.cpp`.
 
 | Ingredient | Available strategies |
 | :---       | :---                 |
-| `QP_solver` | `BQPD` (if available) |
+| `QP_solver` | `BQPD`, `HiGHS` (if available) |
 | `LP_solver` | `BQPD`, `HiGHS` (if available) |
-| `linear_solver` | `MA57`, `MA27`, `MUMPS` (if available) |
+| `linear_solver` | `MA57`, `MA27`, `MUMPS`, `SSIDS` (if available) |
 
 If not provided, the solver is chosen automatically from the available solvers (if any).
 
