### 2026-03-09 3bbc711
Cleaned up the SSIDSSolver class
