### 2026-03-09 3a9bd17
[UnoSolver.jl] Release 0.2.5 (#568)
