### 2026-03-09 d1ff299
[WIP] Interface to SPRAL's SSIDS linear solver (#541)
* First SSIDS interface

* Link SPRAL_jll's libspral in Yggdrasil build + run Julia tests with SSIDS in CI

* Set logger=INFO in uno_ipopt_ssids.jl (most likely SSIDS is not found)

* Set env variables for SPRAL in Yggdrasil build

* Fix the workflows

* Compile SPRAL for the future releases

* Don't compile tests and examples of SPRAL

* Add tests with MUMPS, SSIDS, MA27 and MA57 in UnoSolver.jl

* Fix tests

* Update NLP_wrapper.jl

* Update NLP_wrapper.jl

* Minor change

* Print linear solver in NLPModels tests

* Added SPRAL/SSIDS to t
[…]
