### 2026-03-09 0934dd5
[unopy] Release v0.2.6 (#567)
* [unopy] Test release v0.2.23

* [unopy] Test release v0.2.24

* Bump PyPI version
