### 2026-03-10 95ea827
[UnoSolver.jl] Release fake -- Set OpenMP environment variables in workflow
Added environment variable settings for OpenMP.
