### 2026-03-10 9855b7c
Add a test with HS5
