+++ interfaces/Julia/test/C_wrapper.jl
@@ -1,4 +1,4 @@
-# hs071
+# hs71
 # min x1 * x4 * (x1 + x2 + x3) + x3
 # st  x1 * x2 * x3 * x4 >= 25
 #     x1^2 + x2^2 + x3^2 + x4^2 = 40
@@ -6,18 +6,18 @@
 # Start at (1, 5, 5, 1)
 # End at (1.000..., 4.743..., 3.821..., 1.379...)
 
-function c_objective(userdata::Nothing, x::Vector{Float64})
+function c_objective_hs71(userdata::Nothing, x::Vector{Float64})
   f = x[1] * x[4] * (x[1] + x[2] + x[3]) + x[3]
   return f
 end
 
-function c_constraints(userdata::Nothing, c::Vector{Float64}, x::Vector{Float64})
+function c_constraints_hs71(userdata::Nothing, c::Vector{Float64}, x::Vector{Float64})
   c[1] = x[1] * x[2] * x[3] * x[4]
   c[2] = x[1]^2 + x[2]^2 + x[3]^2 + x[4]^2
   return c
 end
 
-function c_objective_gradient(userdata::Nothing, g::Vector{Float64}, x::Vector{Float64})
+function c_objective_gradient_hs71(userdata::Nothing, g::Vector{Float64}, x::Vector{Float64})
   g[1] = x[1] * x[4] + x[4] * (x[1] + x[2] + x[3])
   g[2] = x[1] * x[4]
   g[3] = x[1] * x[4] + 1
@@ -25,7 +25,7 @@ function c_objective_gradient(userdata::Nothing, g::Vector{Float64}, x::Vector{F
   return g
 end
 
-function c_jacobian(userdata::Nothing, jvals::Vector{Float64}, x::Vector{Float64})
+function c_jacobian_hs71(userdata::Nothing, jvals::Vector{Float64}, x::Vector{Float64})
   # Constraint (row) 1
   jvals[1] = x[2] * x[3] * x[4]  # 1,1
   jvals[2] = x[1] * x[3] * x[4]  # 1,2
@@ -39,7 +39,7 @@ function c_jacobian(userdata::Nothing, jvals::Vector{Float64}, x::Vector{Float64
   return jvals
 end
 
-function c_lagrangian_hessian(
+function c_lagrangian_hessian_hs71(
   userdata::Nothing,
   hvals::Vector{Float64},
   x::Vector{Float64},
@@ -75,7 +75,7 @@ function c_lagrangian_hessian(
   return hvals
 end
 
-function c_jacobian_operator(
+function c_jacobian_operator_hs71(
   userdata::Nothing,
   Jv::Vector{Float64},
   x::Vector{Float64},
@@ -87,7 +87,7 @@ function c_jacobian_operator(
   return Jv
 end
 
-function c_jacobian_transposed_operator(
+function c_jacobian_transposed_operator_hs71(
   userdata::Nothing,
   Jtv::Vector{Float64},
   x::Vector{Float64},
@@ -101,7 +101,7 @@ function c_jacobian_transposed_operator(
   return Jtv
 end
 
-function c_lagrangian_hessian_operator(
+function c_lagrangian_hessian_operator_hs71(
   userdata::Nothing,
   Hv::Vector{Float64},
   x::Vector{Float64},
@@ -173,196 +173,410 @@ function sparsity_pattern_lagrangian_hessian_hs71()
   return hrows, hcols, nnzh
 end
 
-nvar = 4
-lvar = [1.0, 1.0, 1.0, 1.0]
-uvar = [5.0, 5.0, 5.0, 5.0]
[…]
