### 2026-03-10 fc36338
Update build_tarballs_utils.jl for SSIDS
