### 2026-03-10 d4ee43a
Hwloc uses frameworks of Mac... RIP static libary
