### 2026-03-10 35911db
Hwloc uses frameworks of Mac... RIP static libary
