### 2026-03-10 96f013f
Removed catch in uno_ampl, added one in Uno.cpp in case Uno::pick_ingredients() fails
