### 2026-03-10 094b924
Catch std::invalig_argument at startup if one of the options does not exist (#572)
Removed catch in uno_ampl, added one in Uno.cpp in case Uno::pick_ingredients() fails
