+++ interfaces/AMPL/uno_ampl.cpp
@@ -22,80 +22,70 @@ void* operator new(size_t size) {
 
 namespace uno {
    void run_uno_ampl(const std::string& model_name, Options& options) {
-      try {
-         const AMPLModel model(model_name);
-         Uno uno{};
-         Result result = uno.solve(model, options);
-         if (result.optimization_status == OptimizationStatus::SUCCESS) {
-            // check result.solution_status
-         }
-         else {
-            // ...
-         }
-         if (options.get_bool("AMPL_write_solution_to_file")) {
-            model.write_solution_to_file(result);
-         }
-         // std::cout << "memory_allocation_amount = " << memory_allocation_amount << '\n';
+      const AMPLModel model(model_name);
+      Uno uno{};
+      Result result = uno.solve(model, options);
+      if (result.optimization_status == OptimizationStatus::SUCCESS) {
+         // check result.solution_status
       }
-      catch (std::exception& exception) {
-         DISCRETE << exception.what() << '\n';
+      else {
+         // ...
       }
+      if (options.get_bool("AMPL_write_solution_to_file")) {
+         model.write_solution_to_file(result);
+      }
+      // std::cout << "memory_allocation_amount = " << memory_allocation_amount << '\n';
    }
 } // namespace
 
 int main(int argc, char* argv[]) {
    using namespace uno;
 
-   try {
-      if (argc == 1 || (argc == 2 && std::string(argv[1]) == "--v")) {
-         std::cout << "Uno " << Uno::current_version() << '\n';
-      } 
-      else if (argc == 2 && std::string(argv[1]) == "--dump-options") {
-         // Print all available options (type + default value) for automated tools
-         Options::dump_default_options();
-      }
-      else if (argc == 2 && std::string(argv[1]) == "--strategies") {
-         Uno::print_available_strategies();
+   if (argc == 1 || (argc == 2 && std::string(argv[1]) == "--v")) {
+      std::cout << "Uno " << Uno::current_version() << '\n';
+   }
+   else if (argc == 2 && std::string(argv[1]) == "--dump-options") {
+      // Print all available options (type + default value) for automated tools
+      Options::dump_default_options();
+   }
+   else if (argc == 2 && std::string(argv[1]) == "--strategies") {
+      Uno::print_available_strategies();
+   }
+   else {
+      // AMPL expects: ./uno_ampl model.nl [-AMPL] [option_name=option_value, ...]
+      // model name
+      const char* model_name = argv[1];
+
+      // gather the options
+      Options options;
+      DefaultOptions::load(options);
+      // the -AMPL flag indicates that the solution should be written to the AMPL solution file
+      size_t offset = 2;
+      if (argc > 2 && std::string(argv[2]) == "-AMPL") {
+         options.set_bool("AMPL_write_solution_to_file", true);
+         ++offset;
       }
       else {
-         // AMPL expects: ./uno_ampl model.nl [-AMPL] [option_name=option_value, ...]
-         // model name
-         const char* model_name = argv[1];
-
-         // gather the options
-         Options options;
-         DefaultOptions::load(options);
-         // the -AMPL flag indicates that the solution should be written to the AMPL solution file
-         size_t offset = 2;
-         if (argc > 2 && std::string(argv[2]) == "-AMPL") {
-            options.set_bool("AMPL_write_solution_to_file", true);
-            ++offset;
-         }
-         else {
-            options.set_bool("AMPL_write_solution_to_file", false);
-         }
-         // get the command line arguments (options start at index offset)
-         const Options command_line_options = Options::get_command_line_options(argc, argv, offset);
-         // possibly set options from an option file
-         const auto optional_option_file = command_line_options.get_string_optional("option_file");
-         if (optional_option_file.has_value()) {
-            Options file_options = Options::load_option_file(*optional_option_file);
[…]
