### 2026-03-10 c3c5204
[UnoSolver.jl] Release fake -- Set OpenMP environment variables in workflow
