+++ .github/workflows/julia-tests-jll.yml
@@ -37,11 +37,11 @@ jobs:
           echo "OMP_PROC_BIND=TRUE" >> $GITHUB_ENV
 
       - name: Install compilers
+        if: matrix.os == 'macos-latest'
         uses: fortran-lang/setup-fortran@main
         with:
           compiler: 'gcc'
           version: '14'
-        if: matrix.os == 'macos-latest'
 
       - uses: julia-actions/setup-julia@v2
         with:
