### 2026-03-10 cc80469
[UnoSolver.jl] Release fake -- Set OpenMP environment variables in workflow
