### 2026-03-11 76eb7fa
Added parent class QuasiNewtonHessian to existing LBFGSHessian
