### 2026-03-11 3151720
Added parent class QuasiNewtonHessian to existing LBFGSHessian
