+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.cpp
@@ -21,53 +21,17 @@
 
 namespace uno {
    LBFGSHessian::LBFGSHessian(const Model& model, double objective_multiplier, const Options& options):
-         HessianModel("L-BFGS"),
-         model(model),
-         fixed_objective_multiplier(objective_multiplier),
-         memory_size(options.get_unsigned_int("quasi_newton_memory_size")),
+         QuasiNewtonHessian("L-BFGS", model, objective_multiplier, options),
          // matrices
-         S(this->model.number_variables, this->memory_size),
-         Y(this->model.number_variables, this->memory_size),
          L(this->memory_size, this->memory_size),
          D(this->memory_size),
          invsqrt_D(this->memory_size),
          L_invsqrt_D(this->memory_size, this->memory_size),
          M(this->memory_size, this->memory_size),
          U(this->model.number_variables, this->memory_size),
-         V(this->model.number_variables, this->memory_size),
-         current_lagrangian_gradient(this->model.number_variables),
-         trial_lagrangian_gradient(this->model.number_variables) {
+         V(this->model.number_variables, this->memory_size) {
       if (this->memory_size <= 0) {
          throw std::runtime_error("The quasi-Newton memory size should be positive");
-      }
-   }
-
-   bool LBFGSHessian::has_hessian_operator() const {
-      return true;
-   }
-
-   // this will NOT be called by WoodburyEQPSolver
-   bool LBFGSHessian::has_hessian_matrix() const {
-      // never form the explicit, dense matrix
-      return false;
-   }
-
-   bool LBFGSHessian::has_curvature() const {
-      return true;
-   }
-
-   // this can only be called by WoodburyEQPSolver
-   size_t LBFGSHessian::number_nonzeros() const {
-      // count only the diagonal contribution
-      return this->model.number_variables;
-   }
-
-   // this can only be called by WoodburyEQPSolver
-   void LBFGSHessian::compute_sparsity(int* row_indices, int* column_indices, int solver_indexing) const {
-      // diagonal contribution
-      for (size_t variable_index: Range(this->model.number_variables)) {
-         row_indices[variable_index] = static_cast<int>(variable_index) + solver_indexing;
-         column_indices[variable_index] = static_cast<int>(variable_index) + solver_indexing;
       }
    }
 
@@ -186,22 +150,6 @@ namespace uno {
    }
 
    // protected member functions
-
-   void LBFGSHessian::update_S(const Iterate& current_iterate, const Iterate& trial_iterate) {
-      // TODO check that the S entry isn't too small
-      this->S.column(this->current_index) = view(trial_iterate.primals, 0, this->model.number_variables) -
-         view(current_iterate.primals, 0, this->model.number_variables);
-   }
-   
-   // fill the Y matrix: y = \nabla L(x_k, y_k, z_k) - \nabla L(x_{k-1}, y_k, z_k)
-   void LBFGSHessian::update_Y(const Iterate& current_iterate, const Iterate& trial_iterate, EvaluationCache& evaluation_cache) {
-      // evaluate Lagrangian gradients at the current and trial iterates, both with the trial multipliers trial_iterate.multipliers
-      this->model.evaluate_lagrangian_gradient(current_iterate.primals, trial_iterate.multipliers, this->fixed_objective_multiplier,
-         evaluation_cache.current_evaluations, this->current_lagrangian_gradient);
-      this->model.evaluate_lagrangian_gradient(trial_iterate.primals, trial_iterate.multipliers, this->fixed_objective_multiplier,
-         evaluation_cache.trial_evaluations, this->trial_lagrangian_gradient);
[…]

[…]
