### 2026-03-11 b4861ab
Include headers to release artifacts (#576)
Added include to the list of directories in generate_binaries.jl
