+++ .github/julia/generate_binaries.jl
@@ -52,15 +52,15 @@ for (platform, libdir, ext) in platforms
 
       # Create a folder with the version number of the package
       mkdir("$(package)_binaries.$version2")
-      for folder in ("lib", "bin", "share")
+      for folder in ("lib", "bin", "share", "include")
         cp(folder, "$(package)_binaries.$version2/$folder")
       end
 
       cd("$(package)_binaries.$version2")
       if ext == "dll"
-        run(`zip -r --symlinks ../../../$(package)_binaries.$version2.$platform.zip lib bin share`)
+        run(`zip -r --symlinks ../../../$(package)_binaries.$version2.$platform.zip lib bin share include`)
       else
-        run(`tar -czf ../../../$(package)_binaries.$version2.$platform.tar.gz lib bin share`)
+        run(`tar -czf ../../../$(package)_binaries.$version2.$platform.tar.gz lib bin share include`)
       end
       cd("../../..")
 
