+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.cpp
@@ -7,13 +7,11 @@
 #include "linear_algebra/BLAS.hpp"
 #include "linear_algebra/Vector.hpp"
 #include "linear_algebra/VectorView.hpp"
-#include "optimization/EvaluationCache.hpp"
 #include "optimization/Iterate.hpp"
 #include "symbolic/Inverse.hpp"
 #include "symbolic/Multiplication.hpp"
 #include "symbolic/Range.hpp"
 #include "symbolic/ScalarMultiple.hpp"
-#include "symbolic/Subtraction.hpp"
 #include "symbolic/Sum.hpp"
 #include "symbolic/Transpose.hpp"
 #include "tools/Logger.hpp"
