### 2026-03-11 0461150
Added empty class LSR1Hessian + updated HessianModelFactory. If no Hessian matrix/operator provided, factory defaults to L-BFGS for LS and L-SR1 for TR
