### 2026-03-11 fc0a6cc
Temporarily disabled LSR1 in factory
