### 2026-03-11 26bc4e6
Disable fix with dual signs for L-BFGS
