+++ interfaces/Julia/ext/UnoSolverMathOptInterfaceExt/MOI_wrapper.jl
@@ -1676,7 +1676,7 @@ end
 
 function _dual_multiplier(model::Optimizer)
     multiplier = xor(model.problem_type == "LP", model.sense == MOI.MAX_SENSE) ? 1.0 : -1.0
-    multiplier = _quasi_newton_approximation(model) ? -multiplier : multiplier
+    # multiplier = _quasi_newton_approximation(model) ? -multiplier : multiplier
     return multiplier
 end
 
