### 2026-03-11 ed1d3d2
Print problem type in INFO mode
