### 2026-03-11 dab492a
Set logger to INFO in Julia/test/MOI_wrapper.jl
