+++ interfaces/Julia/test/MOI_wrapper.jl
@@ -28,7 +28,7 @@ function test_MOI_Test()
         with_cache_type = Float64,
         with_bridge_type = Float64,
     )
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     MOI.Test.runtests(
         model,
         MOI.Test.Config(
@@ -69,7 +69,7 @@ end
 
 function test_ConstraintDualStart()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     x = MOI.add_variables(model, 2)
     c = MOI.add_constraint(
         model,
@@ -93,7 +93,7 @@ end
 
 function test_ConstraintDualStart_ScalarNonlinearFunction()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     x = MOI.add_variables(model, 2)
     MOI.add_constraint.(model, x, MOI.Interval(0.0, 0.8))
     f = MOI.ScalarNonlinearFunction(:sin, Any[1.0*x[1]-1.0*x[2]])
@@ -115,7 +115,7 @@ end
 
 function test_solve_time()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     MOI.add_variable(model)
     @test isnan(MOI.get(model, MOI.SolveTimeSec()))
     MOI.optimize!(model)
@@ -125,7 +125,7 @@ end
 
 function test_barrier_iterations()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     x = MOI.add_variable(model)
     f = (x - 1.0)^2 + 2.0 * x + 3.0
     MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
@@ -152,7 +152,7 @@ end
 
 function test_empty_optimize()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     @test MOI.get(model, MOI.RawStatusString()) == "Optimize not called"
     MOI.optimize!(model)
     @test MOI.get(model, MOI.TerminationStatus()) == MOI.INVALID_MODEL
@@ -269,7 +269,7 @@ end
 
 function test_scalar_nonlinear_function_nlp_block()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     x = MOI.add_variable(model)
     f = MOI.ScalarNonlinearFunction(:^, Any[x, 4])
     MOI.add_constraint(model, f, MOI.LessThan(1.0))
@@ -282,7 +282,7 @@ end
 
 function test_parameter()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     p, ci = MOI.add_constrained_variable(model, MOI.Parameter(1.0))
     x = MOI.add_variable(model)
     fi = MOI.ScalarNonlinearFunction(:-, Any[x, p])
@@ -303,7 +303,7 @@ end
 
 function test_parameter_replace_parameters()
     model = UnoSolver.Optimizer()
-    MOI.set(model, MOI.Silent(), true)
+    # MOI.set(model, MOI.Silent(), true)
     p, ci = MOI.add_constrained_variable(model, MOI.Parameter(1.0))
[…]
