### 2026-03-11 a69b7a8
Added debug info in Julia/ext/UnoSolverMathOptInterfaceExt/MOI_wrapper.jl
