### 2026-03-11 e4bd9c5
Systematically set Hessian and Hessian operator if available. Let Uno figure out the problem type based on nnzh
