### 2026-03-11 af858af
Allow nnzh = 0 in uno_set_lagrangian_hessian
