### 2026-03-12 137743c
Backport of Alexis' commit
