### 2026-03-12 64a8920
Put the tests back in silent mode
