### 2026-03-12 d3ee329
[C/UnoSolver.jl] Properly set model properties for LPs (#578)
* Disable fix with dual signs for L-BFGS

* Print problem type in INFO mode

* Set logger to INFO in Julia/test/MOI_wrapper.jl

* Added debug info in Julia/ext/UnoSolverMathOptInterfaceExt/MOI_wrapper.jl

* Systematically set Hessian and Hessian operator if available. Let Uno figure out the problem type based on nnzh

* Allow nnzh = 0 in uno_set_lagrangian_hessian

* Backport of Alexis' commit

* Put the tests back in silent mode
