+++ uno/optimization/ProblemType.hpp
@@ -6,6 +6,18 @@
 
 namespace uno {
    enum class ProblemType{LINEAR, QUADRATIC, NONLINEAR};
+
+   inline std::string to_string(ProblemType problem_type) {
+      if (problem_type == ProblemType::LINEAR) {
+         return "LP";
+      }
+      else if (problem_type == ProblemType::QUADRATIC) {
+         return "QP";
+      }
+      else {
+         return "NLP";
+      }
+   }
 } // namespace
 
 #endif // UNO_PROBLEMTYPE_H
