+++ uno/model/Model.hpp
@@ -6,6 +6,8 @@
 
 #include <string>
 #include <vector>
+
+#include "../../interfaces/C/Uno_C_API.h"
 #include "linear_algebra/MatrixOrder.hpp"
 #include "linear_algebra/Norm.hpp"
 #include "optimization/ProblemType.hpp"
@@ -36,7 +38,7 @@ namespace uno {
       const size_t number_variables; /*!< Number of variables */
       const size_t number_constraints; /*!< Number of constraints */
       const double optimization_sense; /*!< 1: minimization, -1: maximization */
-      const double lagrangian_sign_convention; /*!< -1, 1 */
+      double lagrangian_sign_convention;
 
       [[nodiscard]] virtual ProblemType get_problem_type() const = 0;
 
