### 2026-03-12 a2c5b8a
Fixed .github/C/tests_hs015.c
