+++ .github/C/tests_hs015.c
@@ -115,28 +115,25 @@ void solve_instance(uno_int optimization_sense, double lagrangian_sign_conventio
 	assert(uno_set_constraints(model, number_constraints, constraint_functions,
 		constraints_lower_bounds, constraints_upper_bounds, number_jacobian_nonzeros,
 		jacobian_row_indices, jacobian_column_indices, constraint_jacobian));
+	assert(uno_set_lagrangian_convention(model, lagrangian_sign_convention));
 	if (lagrangian_sign_convention == UNO_MULTIPLIER_NEGATIVE) {
 		if (optimization_sense == UNO_MINIMIZE) {
 			assert(uno_set_lagrangian_hessian(model, number_hessian_nonzeros, hessian_triangular_part,
-				hessian_row_indices, hessian_column_indices, lagrangian_hessian_negative_sign,
-				lagrangian_sign_convention));
+				hessian_row_indices, hessian_column_indices, lagrangian_hessian_negative_sign));
 		}
 		else {
 			assert(uno_set_lagrangian_hessian(model, number_hessian_nonzeros, hessian_triangular_part,
-				hessian_row_indices, hessian_column_indices, lagrangian_hessian_max_negative_sign,
-				lagrangian_sign_convention));
+				hessian_row_indices, hessian_column_indices, lagrangian_hessian_max_negative_sign));
 		}
 	}
 	else {
 		if (optimization_sense == UNO_MINIMIZE) {
 			assert(uno_set_lagrangian_hessian(model, number_hessian_nonzeros, hessian_triangular_part,
-				hessian_row_indices, hessian_column_indices, lagrangian_hessian_positive_sign,
-				lagrangian_sign_convention));
+				hessian_row_indices, hessian_column_indices, lagrangian_hessian_positive_sign));
 		}
 		else {
 			assert(uno_set_lagrangian_hessian(model, number_hessian_nonzeros, hessian_triangular_part,
-				hessian_row_indices, hessian_column_indices, lagrangian_hessian_max_positive_sign,
-				lagrangian_sign_convention));
+				hessian_row_indices, hessian_column_indices, lagrangian_hessian_max_positive_sign));
 		}
 	}
 	assert(uno_set_initial_primal_iterate(model, x0));
