+++ uno/model/Model.hpp
@@ -6,13 +6,12 @@
 
 #include <string>
 #include <vector>
-
-#include "../../interfaces/C/Uno_C_API.h"
 #include "linear_algebra/MatrixOrder.hpp"
 #include "linear_algebra/Norm.hpp"
 #include "optimization/ProblemType.hpp"
 #include "symbolic/VectorExpression.hpp"
 #include "../interfaces/C/uno_int.h"
+#include "../interfaces/C/Uno_C_API.h"
 
 namespace uno {
    // forward declarations
