### 2026-03-12 0a31e34
MOI_wrapper.jl: removed dependency on L-BFGS in _dual_multiplier()
