+++ interfaces/Julia/ext/UnoSolverMathOptInterfaceExt/MOI_wrapper.jl
@@ -1679,7 +1679,8 @@ function _quasi_newton_approximation(model::Optimizer)::Bool
 end
 
 function _dual_multiplier(model::Optimizer)
-    multiplier = xor(_quasi_newton_approximation(model), model.sense == MOI.MAX_SENSE) ? 1.0 : -1.0
+    #multiplier = xor(_quasi_newton_approximation(model), model.sense == MOI.MAX_SENSE) ? 1.0 : -1.0
+    multiplier = (model.sense == MOI.MAX_SENSE) ? 1.0 : -1.0
     return multiplier
 end
 
