### 2026-03-12 432ccfa
Added LDL'/signed Cholesky ldlt_nopiv_lvl2_rightlooking() in LAPACK_extension.hpp + mutable view around T*
