### 2026-03-12 6b2bae9
Added LDL'/signed Cholesky ldlt_nopiv_lvl2_rightlooking() in LAPACK_extension.hpp + mutable view around T*
