+++ uno/linear_algebra/VectorView.hpp
@@ -17,10 +17,7 @@ namespace uno {
       using value_type = typename std::remove_reference_t<Vector>::value_type;
 
       VectorView(const Vector& vector, size_t start, size_t end):
-            BLASVector<value_type>(),
-            vector(vector),
-            start(start),
-            end(std::min(end, this->vector.size())) {
+            BLASVector<value_type>(), vector(vector), start(start), end(std::min(end, this->vector.size())) {
          if (this->end < this->start) {
             throw std::runtime_error("The view ends before its starting point");
          }
@@ -51,10 +48,7 @@ namespace uno {
       using value_type = typename std::remove_reference_t<Vector>::value_type;
 
       MutableVectorView(Vector& vector, size_t start, size_t end):
-            MutableBLASVector<value_type>(),
-            vector(vector),
-            start(start),
-            end(std::min(end, this->vector.size())) {
+            MutableBLASVector<value_type>(), vector(vector), start(start), end(std::min(end, this->vector.size())) {
          if (this->end < this->start) {
             throw std::runtime_error("The view ends before its starting point");
          }
@@ -91,6 +85,50 @@ namespace uno {
       const size_t end;
    };
 
+   // span of a C array: allocation-free view of a certain length
+   template <typename T>
+   class MutableArrayView: public MutableBLASVector<T> {
+   public:
+      using value_type = T;
+
+      MutableArrayView(T* array, size_t start, size_t end):
+            MutableBLASVector<value_type>(), array(array), start(start), end(end) {
+         if (this->end < this->start) {
+            throw std::runtime_error("The view ends before its starting point");
+         }
+      }
+
+      // operators
+      using MutableBLASVector<value_type>::operator=;
+      using MutableBLASVector<value_type>::operator+=;
+      using MutableBLASVector<value_type>::operator-=;
+
+      [[nodiscard]] size_t size() const noexcept override {
+         return this->end - this->start;
+      }
+
+      [[nodiscard]] value_type* data() noexcept override {
+         return this->array + this->start;
+      }
+
+      [[nodiscard]] const value_type* data() const noexcept override {
+         return this->array + this->start;
+      }
+
+      [[nodiscard]] value_type& operator[](size_t index) noexcept override {
+         return this->array[this->start + index];
+      }
+
+      [[nodiscard]] const value_type& operator[](size_t index) const noexcept override {
+         return this->array[this->start + index];
+      }
+
+   protected:
+      T* array;
+      const size_t start;
+      const size_t end;
+   };
+
    // free functions
    template <typename Expression>
    auto view(const Expression& expression, size_t start, size_t end) {
@@ -100,6 +138,11 @@ namespace uno {
    template <typename Expression>
    auto view(Expression& expression, size_t start, size_t end) {
       return MutableVectorView{expression, start, end};
+   }
+
+   template <typename T>
+   auto view(T* array, size_t start, size_t end) {
+      return MutableArrayView{array, start, end};
    }
 
    template <typename Expression>
