+++ uno/linear_algebra/BLASMatrix.hpp
@@ -35,11 +35,11 @@ namespace uno {
       template <typename Matrix1, typename Matrix2>
       BLASMatrix<T>& operator+=(Multiplication<Transpose<Matrix1>, Matrix2>&& expression);
 
-      // specialized operator+= for low-rank update C := A A^T
+      // specialized operator+= for rank-k update C := A A^T
       template <typename Matrix>
       BLASMatrix& operator=(Multiplication<Matrix, Transpose<Matrix>>&& expression);
 
-      // specialized operator+= for low-rank update C += alpha A^T A
+      // specialized operator+= for rank-k update C += alpha A^T A
       template <typename Matrix>
       BLASMatrix& operator+=(ScalarMultiple<Multiplication<Transpose<Matrix>, Matrix>>&& expression);
 
@@ -108,7 +108,7 @@ namespace uno {
       return *this;
    }
 
-   // specialized operator+= for low-rank update C := A A^T
+   // specialized operator+= for rank-k update C := A A^T
    template <typename T>
    template <typename Matrix>
    BLASMatrix<T>& BLASMatrix<T>::operator=(Multiplication<Matrix, Transpose<Matrix>>&& expression) {
@@ -120,14 +120,14 @@ namespace uno {
       DEBUG << "Performing rank " << correction_rank << " update\n";
       // check that A and B are the same object
       if (&A != &B) {
-         throw std::runtime_error("BLASMatrix::operator+=: low-rank update called on two different correction matrices");
+         throw std::runtime_error("BLASMatrix::operator+=: rank-k update called on two different correction matrices");
       }
       blas3::symmetric_rank_k_update('L', 'N', this->number_rows, A.number_rows, A.number_columns, 1., A.data(),
          A.leading_dimension, 0., this->data(), this->leading_dimension);
       return *this;
    }
 
-   // specialized operator+= for low-rank update C += alpha A^T A
+   // specialized operator+= for rank-k update C += alpha A^T A
    template <typename T>
    template <typename Matrix>
    BLASMatrix<T>& BLASMatrix<T>::operator+=(ScalarMultiple<Multiplication<Transpose<Matrix>, Matrix>>&& expression) {
@@ -140,7 +140,7 @@ namespace uno {
       DEBUG << "Performing rank " << correction_rank << " update\n";
       // check that A and B are the same object
       if (&A != &B) {
-         throw std::runtime_error("BLASMatrix::operator+=: low-rank update called on two different correction matrices");
+         throw std::runtime_error("BLASMatrix::operator+=: rank-k update called on two different correction matrices");
       }
       blas3::symmetric_rank_k_update('L', 'T', this->number_rows, A.number_rows, A.number_columns, alpha, A.data(),
          A.leading_dimension, 1., this->data(), this->leading_dimension);
