### 2026-03-12 31ec754
Alexis suggestions: renamed uno_set_lagrangian_convention() to uno_set_lagrangian_sign_convention(), documented default value, and updated unopy README
