+++ interfaces/Julia/src/libuno.jl
@@ -98,9 +98,9 @@ function uno_set_lagrangian_hessian_operator(model, lagrangian_hessian_operator)
                                                       lagrangian_hessian_operator::Ptr{Cvoid})::Bool
 end
 
-function uno_set_lagrangian_convention(model, lagrangian_sign_convention)
-    @ccall libuno.uno_set_lagrangian_convention(model::Ptr{Cvoid},
-                                                lagrangian_sign_convention::Cdouble)::Bool
+function uno_set_lagrangian_sign_convention(model, lagrangian_sign_convention)
+    @ccall libuno.uno_set_lagrangian_sign_convention(model::Ptr{Cvoid},
+                                                     lagrangian_sign_convention::Cdouble)::Bool
 end
 
 function uno_set_user_data(model, user_data)
