+++ CMakeLists.txt
@@ -378,5 +378,5 @@ install(TARGETS ${INSTALL_TARGETS}
    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}  # DLLs on windows
    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}  # static libraries
    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}  # shared libraries
-   PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR} # headers
+   PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/uno # headers
 )
