### 2026-03-12 caafc53
Added include(GNUInstallDirs) to CMake. Required for installing wheels
