### 2026-03-12 574de8c
Install headers in `uno` subdirectory of include (#581)
* Install headers in ${CMAKE_INSTALL_INCLUDEDIR}/uno

* Fixed C examples (use #include "uno/Uno_C_API.h")

* Added include(GNUInstallDirs) to CMake. Required for installing wheels
