+++ CMakeLists.txt
@@ -372,11 +372,12 @@ endif()
 if(TARGET uno_ampl)
    list(APPEND INSTALL_TARGETS uno_ampl)
 endif()
+include(GNUInstallDirs)  # ensures CMAKE_INSTALL_INCLUDEDIR is set to "include"
 install(TARGETS ${INSTALL_TARGETS}
    COMPONENT libuno
    EXPORT UnoTargets
    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}  # DLLs on windows
    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}  # static libraries
    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}  # shared libraries
-   PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR} # headers
+   PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/uno # headers
 )
