### 2026-03-12 e568394
[UnoSolver.jl] Release 0.2.6
