+++ interfaces/Julia/Project.toml
@@ -1,7 +1,7 @@
 name = "UnoSolver"
 uuid = "1baa60ac-02f7-4b39-a7a8-2f4f58486b05"
 authors = ["Alexis Montoison <amontoison@anl.gov>", "Charlie Vanaret <vanaret@zib.de>"]
-version = "0.2.5"
+version = "0.2.6"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -22,7 +22,7 @@ MathOptInterface = "1.48"
 NLPModels = "0.21.10"
 OpenBLAS32_jll = "0.3.24"
 Test = "1.10"
-Uno_jll = "=2.4.1"
+Uno_jll = "=2.4.2"
 julia = "1.10"
 
 [extras]
