### 2026-03-12 8c6d86f
Backported Alexis' commit
