### 2026-03-12 1c8d173
[UnoSolver.jl] Release 0.2.6 (#586)
* [UnoSolver.jl] Release 0.2.6

* Backported Alexis' commit

* Update julia-tests-jll.yml

---------

Co-authored-by: Alexis Montoison <35051714+amontoison@users.noreply.github.com>
