+++ interfaces/Julia/gen/Project.toml
@@ -5,4 +5,4 @@ Uno_jll = "396d5378-14f1-5ab1-981d-48acd51740ed"
 
 [compat]
 julia = "1.10"
-Uno_jll = "=2.4.1"
+Uno_jll = "=2.4.2"
