### 2026-03-13 7addf1d
[CI] Compile Uno with SPRAL
