+++ dependencies/scripts/download_dependencies.sh
@@ -32,10 +32,10 @@ curl -L -o BQPD.tar.gz "$ASSET_URL"
 tar -xzvf BQPD.tar.gz
 pwd
 
-# download UnoUtils: MUMPS (+ METIS, BLAS and LAPACK) and HiGHS
-VERSION="2026.2.17"
+# download UnoUtils: MUMPS (+ METIS, BLAS and LAPACK), HiGHS and SPRAL
+VERSION="2026.3.10"
 REPO="https://github.com/amontoison/UnoUtils_jll.jl/releases/download/UnoUtils-v${VERSION}%2B0"
-ASSET_NAME="UnoUtils.v${VERSION}.${ARCH}-${OS}-libgfortran5.tar.gz"
+ASSET_NAME="UnoUtils.v${VERSION}.${ARCH}-${OS}-libgfortran5-cxx11.tar.gz"
 ASSET_URL="${REPO}/${ASSET_NAME}"
 echo "Downloading: ${ASSET_URL}"
 curl -L -o UnoUtils.tar.gz "$ASSET_URL"
@@ -45,11 +45,3 @@ pwd
 # delete unwanted directories
 # rm -rf lib/cmake/cblas* lib/cmake/lapack* lib/pkgconfig
 rm -rf lib/cmake lib/pkgconfig
-
-if [[ "$OS" == "w64-mingw32" ]]; then
-    cd ..
-    ASSET_URL="https://github.com/JuliaLang/PackageCompiler.jl/releases/download/v1.0.0/x86_64-8.1.0-release-posix-seh-rt_v6-rev0.tar.gz"
-    curl -L -o libstdc++.tar.gz "$ASSET_URL"
-    tar -xzvf libstdc++.tar.gz
-    cp ./mingw64/lib/gcc/x86_64-w64-mingw32/8.1.0/libstdc++.a ./dependencies/lib/libstdc++.a
-fi
