### 2026-03-13 22120bd
[CI] Compile Uno with SPRAL
