+++ .github/workflows/C-test-example.yml
@@ -58,7 +58,7 @@ jobs:
                                                -DHIGHS=${{github.workspace}}/dependencies/lib/libhighs.a \
                                                -DSPRAL=${{github.workspace}}/dependencies/lib/libspral.a \
                                                -DLAPACK_LIBRARIES=${{github.workspace}}/dependencies/lib/liblapack.a \
-                                               -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a;${{github.workspace}}/dependencies/lib/libhwloc.a" \
+                                               -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a;${{github.workspace}}/dependencies/lib/libhwloc.so" \
                                                -DBUILD_STATIC_LIBS=OFF \
                                                -DBUILD_SHARED_LIBS=ON .
 
@@ -78,7 +78,7 @@ jobs:
                 -DCMAKE_BUILD_TYPE=${{env.BUILD_TYPE}} \
                 -DUNO=/usr/local/lib/libuno.so \
                 -DLAPACK_LIBRARIES=${{github.workspace}}/dependencies/lib/liblapack.a \
-                -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libblas.a" .
+                -DBLAS_LIBRARIES=${{github.workspace}}/dependencies/lib/libblas.a .
 
       - name: Build example
         working-directory: ${{github.workspace}}/interfaces/C/example
