### 2026-03-13 0639213
Fix download_dependencies.sh
