+++ .github/workflows/unit-tests.yml
@@ -112,7 +112,7 @@ jobs:
               -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a" \
               -DLAPACK_LIBRARIES=${{github.workspace}}/dependencies/lib/liblapack.a \
               -DHIGHS=${{github.workspace}}/dependencies/lib/libhighs.a \
-              -DHIGHS=${{github.workspace}}/dependencies/lib/libspral.a \
+              -DSPRAL=${{github.workspace}}/dependencies/lib/libspral.a \
               -DBUILD_STATIC_LIBS=ON \
               -DBUILD_SHARED_LIBS=OFF .
 
