### 2026-03-13 81a842c
Passed libhwloc to example CMake in C-test-example.yml
