### 2026-03-13 8b584f4
Added libhwloc to tests CMake in C-tests-linux.yml
