### 2026-03-13 b0364ad
Introduced CMake variables AUXILIARY_VARIABLES + adapted build.yml and build-python-wheels.yml to link libstdc++
