### 2026-03-13 5b3fc05
Link against AUXILIARY_VARIABLES with absolute paths
