+++ pyproject.toml
@@ -8,9 +8,9 @@ build-backend = "scikit_build_core.build"
 [project]
 name = "unopy"
 # PyPI
-#version = "0.2.6"
+version = "0.2.7"
 # TestPyPI
-version = "0.2.25"
+#version = "0.2.25"
 description = "With Uno, finally take full control of your SQP/barrier solver for nonlinearly constrained optimization"
 readme = "interfaces/Python/README.md"
 authors = [
