### 2026-03-13 9367e24
[unopy] Release v0.2.7 (#587)
* [unopy] Test release v0.2.25

* Bumped PyPI version
