### 2026-03-13 c10fd1b
Introduced CMake variable `AUXILIARY_LIBRARIES` (#589)
* Introduced CMake variables AUXILIARY_VARIABLES + adapted build.yml and build-python-wheels.yml to link libstdc++

* Link against AUXILIARY_VARIABLES with absolute paths
