+++ CMakeLists.txt
@@ -96,7 +96,10 @@ file(GLOB TESTS_UNO_SOURCE_FILES
 #########################
 # external dependencies #
 #########################
+# list of all libraries
 set(LIBRARIES "")
+# additional libraries to link against
+set(AUXILIARY_LIBRARIES "" CACHE STRING "Additional libraries to link against")
 
 # function that links an existing library to Uno
 function(link_to_uno library_name library_path)
@@ -239,6 +242,16 @@ if(BLAS_LIBRARIES)
 else()
    find_package(BLAS REQUIRED)
    list(APPEND LIBRARIES BLAS::BLAS)
+endif()
+
+# add the auxiliary libraries last
+if(AUXILIARY_LIBRARIES)
+   set(ABSOLUTE_AUXILIARY_LIBRARIES)
+   foreach(relative_path IN LISTS AUXILIARY_LIBRARIES)
+      get_filename_component(absolute_path "${relative_path}" ABSOLUTE)
+      list(APPEND ABSOLUTE_AUXILIARY_LIBRARIES "${absolute_path}")
+   endforeach()
+   list(APPEND LIBRARIES ${ABSOLUTE_AUXILIARY_LIBRARIES})
 endif()
 
 ###############
