### 2026-03-13 23d25cd
Moved libhwloc to AUXLIARY_LIBRARIES for Uno build
