+++ .github/workflows/build.yml
@@ -180,10 +180,11 @@ jobs:
           -DMUMPS_COMMON_LIBRARY=${{github.workspace}}/dependencies/lib/libmumps_common.a \
           -DMUMPS_PORD_LIBRARY=${{github.workspace}}/dependencies/lib/libpord.a \
           -DMUMPS_MPISEQ_LIBRARY=${{github.workspace}}/dependencies/lib/libmpiseq.a \
-          -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a;${{github.workspace}}/dependencies/lib/libhwloc.so" \
+          -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a" \
           -DLAPACK_LIBRARIES=${{github.workspace}}/dependencies/lib/liblapack.a \
           -DHIGHS=${{github.workspace}}/dependencies/lib/libhighs.a \
           -DSPRAL=${{github.workspace}}/dependencies/lib/libspral.a \
+          -DAUXILIARY_LIBRARIES="${{github.workspace}}/dependencies/lib/libhwloc.so" \
           -DBUILD_STATIC_LIBS=ON \
           -DBUILD_SHARED_LIBS=ON
 
@@ -210,10 +211,11 @@ jobs:
           -DMUMPS_COMMON_LIBRARY=${{github.workspace}}/dependencies/lib/libmumps_common.a \
           -DMUMPS_PORD_LIBRARY=${{github.workspace}}/dependencies/lib/libpord.a \
           -DMUMPS_MPISEQ_LIBRARY=${{github.workspace}}/dependencies/lib/libmpiseq.a \
-          -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a;${{github.workspace}}/dependencies/lib/libhwloc.dylib" \
+          -DBLAS_LIBRARIES="${{github.workspace}}/dependencies/lib/libcblas.a;${{github.workspace}}/dependencies/lib/libblas.a" \
           -DLAPACK_LIBRARIES=${{github.workspace}}/dependencies/lib/liblapack.a \
           -DHIGHS=${{github.workspace}}/dependencies/lib/libhighs.a \
           -DSPRAL=${{github.workspace}}/dependencies/lib/libspral.a \
+          -DAUXILIARY_LIBRARIES="${{github.workspace}}/dependencies/lib/libhwloc.dylib" \
           -DBUILD_STATIC_LIBS=ON \
           -DBUILD_SHARED_LIBS=ON
 
