### 2026-03-13 c30cf17
Upgraded CMake files of C example and C tests with AUXILIARY_LIBRARIES + adapted the workflows
