### 2026-03-13 c68a886
ldlt_nopiv_lvl2_rightlooking() implementation without view()
