### 2026-03-13 3686ed2
ldlt_nopiv_lvl2_rightlooking() implementation without view()
