+++ uno/ingredients/hessian_models/quasi_newton/LBFGSHessian.cpp
@@ -42,23 +42,15 @@ namespace uno {
    void LBFGSHessian::notify_accepted_iterate(Statistics& statistics, const Iterate& current_iterate, const Iterate& trial_iterate,
          EvaluationCache& evaluation_cache) {
       DEBUG << "\n*** Updating the BFGS memory at slot " << this->current_index << '\n';
-      // update the matrices S, Y and D
-      this->update_S(current_iterate, trial_iterate);
-      this->update_Y(current_iterate, trial_iterate, evaluation_cache);
-      this->update_D();
-      DEBUG << "> S: " << this->S;
-      DEBUG << "> Y: " << this->Y;
-      DEBUG << "> diag(D): "; print_vector(DEBUG, this->D);
+      // update the matrices S and Y
+      this->update_limited_memory(current_iterate, trial_iterate, evaluation_cache);
 
-      // check that the latest D entry s^T y is > 0
+      // check that the latest D entry sᵀ y is > 0
       // TODO implement Procedure 18.2 (Damped BFGS Updating) from Numerical Optimization
+      this->update_D();
+      DEBUG << "> diag(D): "; print_vector(DEBUG, this->D);
       if (0. < this->D[this->current_index]) {
-         DEBUG << "S, Y and D updated at slot " << this->current_index << '\n';
-         this->number_entries_in_memory = std::min(this->number_entries_in_memory + 1, this->memory_size);
-         // notify_accepted_iterate is called at the end of a major iteration. Since we don't know yet whether the L-BFGS
-         // Hessian will be used, we delay the update to the beginning of the next major iteration
-         this->hessian_recomputation_required = true;
-         DEBUG << "There are now " << this->number_entries_in_memory << " entries in memory (capacity " << this->memory_size << ")\n";
+         this->validate_update();
       }
       else {
          DEBUG << "Skipping the update\n";
@@ -66,12 +58,12 @@ namespace uno {
       statistics.set("|BFGS|", this->number_entries_in_memory);
    }
 
-   // Hessian-vector product where the Hessian approximation is Bk = B0 - U U^T + V V^T and B0 = delta I
-   // Bk v = (B0 - U U^T + V V^T) v = delta v - U (U^T v) + V (V^T v)
+   // Hessian-vector product where the Hessian approximation is Bk = B0 - U Uᵀ + V Vᵀ and B0 = delta I
+   // Bk v = (B0 - U Uᵀ + V Vᵀ) v = delta v - U (Uᵀ v) + V (Vᵀ v)
    void LBFGSHessian::compute_hessian_vector_product(const double* /*x*/, const double* vector,
          double objective_multiplier, const Vector<double>& /*constraint_multipliers*/, double* result) {
       if (objective_multiplier != this->fixed_objective_multiplier) {
-         throw std::runtime_error("The L-BFGS Hessian model was initialized with a different objective multiplier");
+         throw std::runtime_error("The quasi-Newton Hessian model was initialized with a different objective multiplier");
       }
 
       // update_limited_memory() has updated the limited memory. A recomputation of the Hessian representation may be required
@@ -87,20 +79,20 @@ namespace uno {
 
       // rank-2 contribution
       // (U, V) in R^{n x m}
-      int n = static_cast<int>(this->model.number_variables);
-      int increment = 1;
-      // U U^T v
+      const int n = static_cast<int>(this->model.number_variables);
+      constexpr int increment = 1;
+      // work on each column of U (Uᵀ v)
       for (size_t column_index: Range(this->number_entries_in_memory)) {
[…]
