+++ uno/ingredients/hessian_models/quasi_newton/QuasiNewtonHessian.hpp
@@ -36,7 +36,10 @@ namespace uno {
       DenseMatrix<double> Y;
       Vector<double> current_lagrangian_gradient;
       Vector<double> trial_lagrangian_gradient;
+      bool hessian_recomputation_required{false};
 
+      void update_limited_memory(const Iterate& current_iterate, const Iterate& trial_iterate, EvaluationCache& evaluation_cache);
+      void validate_update();
       void update_S(const Iterate& current_iterate, const Iterate& trial_iterate);
       void update_Y(const Iterate& current_iterate, const Iterate& trial_iterate, EvaluationCache& evaluation_cache);
    };
