### 2026-03-14 e76e8eb
Decrease TR_min_radius from 1e-7 to 1e-12
