### 2026-03-15 cde645e
Restored lagrangian_sign in unconstrained model creation (as proposed in https://github.com/cvanaret/Uno/pull/561). My bad!
