### 2026-03-15 1944e20
Pass Lagrangian sign convention to model creation in a separate function (#580)
* Added function uno_set_lagrangian_convention() to set Lagrangian sign convention, removed it from uno_set_lagrangian_hessian() and uno_set_lagrangian_hessian_operator()

* Fixed Fortran example

* Fixed .github/C/tests_hs015.c

* Minor change

* Minor change

* MOI_wrapper.jl: removed dependency on L-BFGS in _dual_multiplier()

* Cleanup

* Alexis suggestions: renamed uno_set_lagrangian_convention() to uno_set_lagrangian_sign_convention(), documented default value, and updated unopy README

* Restored lagrangian_
[…]
