### 2026-03-16 04ed032
Make UNO_MULTIPLIER_POSITIVE and UNO_MULTIPLIER_NEGATIVE integers
