### 2026-03-16 957b7aa
Make `UNO_MULTIPLIER_POSITIVE` and `UNO_MULTIPLIER_NEGATIVE` integers (#593)
* Make UNO_MULTIPLIER_POSITIVE and UNO_MULTIPLIER_NEGATIVE integers

* Regenerated libuno.jl

* Fixed forgotten changes
