+++ uno/model/Model.cpp
@@ -15,7 +15,7 @@
 namespace uno {
    // abstract Problem class
    Model::Model(std::string name, size_t number_variables, size_t number_constraints, double objective_sign,
-      double lagrangian_sign_convention) :
+                double lagrangian_sign_convention) :
          name(std::move(name)), number_variables(number_variables), number_constraints(number_constraints),
          optimization_sense(objective_sign), lagrangian_sign_convention(lagrangian_sign_convention) {
    }
