### 2026-03-18 #10229 @editorialbot
[REVIEW]: Uno: a composable framework for nonlinearly constrained optimization
**Submitting author:** @cvanaret (<a href="http://orcid.org/0000-0002-1131-7631">Charlie Vanaret</a>)
**Repository:** https://github.com/cvanaret/Uno
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.3.1
