### 2026-04-21 #10229 @victoraalves
@cvanaret Great tool. I recommend this for publication after a few, (very minor IMHO) edits:

Major action points:
1. **Installation.** I'd be nice to have more streamlined installation options. I'd move the precompiled/executable instructions to the top of [INSTALL.md](https://github.com/cvanaret/Uno/blob/main/INSTALL.md). I'd also streamline the use of unopy for Python users.
2. **Documentation.** Currently, installation instructions, instructions on using the different interfaces, case-studies/examples and theory are scattered across multiple *.md files cited in the main README.md in the project's root folder as well as in https://unosolver.readthedocs.io/en/latest/. I'd recommend unifying **all** of this information in a single readthedocs webpage. This will centralize installation instructions, theory, examples and so on. You already have everything here, it's just about reorganizing information. I've found a bit confusing to navigate throughout several *.md files and the page to understand all aspects of Uno.
3. **OS specific instructions.** The "Precompiled libraries and executables" section on [INSTALL.md](https://github.com/cvanaret/Uno/blob/main/INSTALL.md) gives somewhat generic/somewhat linux-ish instructions. I'd recommend having OS-specific sections and move this to the top of this page (per comment 1 above).

Minor action points:

1. While restructuring your documentation, I have a minor suggestion: Having a section quickly showing uno through the interfaces. Something like a "Getting Started" or "Usage". [One example can be seen here.](https://cyipopt.readthedocs.io/stable/tutorial.html). Maybe showing unopy + other interfaces in action. I was able to run [example_hs015.py](https://github.com/cvanaret/Uno/blob/main/interfaces/Python/example/example_hs015.py) successfully. Maybe this is a good candidate?

Once the major action points are taken care of, I feel like I can check above the _**"Installation instructions"**_ item.
All in all, strong submission. I recommend this for publication. @drvinceknight.
