### 2026-05-12 #10229 @anugrahjo
Paper review:

The software appears promising and useful for optimization researchers and solver developers. However, the paper needs significant revisions before it can be accepted. Currently, the paper feels like a collection of descriptions rather than a coherent argument. The current manuscript lacks a clear narrative, the motivation is underdeveloped, and the relationship between Uno and existing optimization software is not sufficiently established.

The **statement of need** is currently too generic. It explains that nonlinear optimization solvers share common components, but it does not clearly mention the specific gap Uno fills. The paper says algorithmic ideas are usually tested at the level of complete solvers rather than individual components, motivating a composable framework, but this point is not developed enough to establish why existing tools are insufficient or why Uno is needed. At present, the motivation reads more like a broad observation about nonlinear optimization than a compelling case for this particular software. A statement of need should answer: Who is the target audience: solver developers, optimization researchers, modeling-language users, or end users?, What is difficult or impossible to do with existing solvers/frameworks?.

The **state-of-the-field** section is very short: it only briefly discusses IPOPT, KNITRO, and SNOPT. Since Uno’s main contribution is composability, the paper must distinguish itself from existing composable or modular optimization frameworks, not only from monolithic production solvers. Related tools already exist, e.g., [modOpt](https://www.sciencedirect.com/science/article/abs/pii/S0965997825002224) has similar goals and is designed to support fast development and testing of optimization algorithms using self-contained modules. Additionally, other nonlinear optimization libraries such as scipy.optimize, pyopt, pyoptsparse, nlopt, etc., do allow customizing their code, however, not via modular components. Please add a substantive comparison to these solvers, rather than just citing them. Without discussing and differentiating Uno from them, the novelty and need for Uno are not adequately established.

The organization could be substantially improved. The **Unification framework** section is too short to be a separate section, as the idea in its first paragraph was already described in the **Summary** section. Consider merging it with **Software design** or **Summary**.  When describing the "wheel of strategies", it won't be clear to a non-expert audience what the authors mean by "high-level layers", "ingredients", and "strategies" without a bit more explanation.

In the **Software design** section L60, "agnostic to" will sound more natural than "agnostic of".

The **Interfaces** section is particularly unfocused and too long relative to the rest of the paper and reads like implementation/user documentation rather than a JOSS paper. It provides detailed descriptions of the AMPL, C, Fortran, Julia, and Python interfaces, including file names such as `uno_c.f90` and `uno_fortran.f90`, opaque C objects, wrapper details, and distribution mechanisms. This section should be condensed and reframed around the user workflows enabled by each interface. Instead of describing each interface in paragraph form (some paragraphs contain just one sentence), consider using a compact table or bulleted list.

The **Research impact statement** also needs major revisions. The research impact statement mixes several types of claims: benchmark performance, conference presentations, invited talks, collaborations, current use in external packages, and possible future integrations. Some of this material is appropriate, but the section needs better organization. Benchmark claims should be placed in a different section, maybe in the software design section, right after the paragraph that talks about presets. The paragraph listing conference presentations and invited talks is weaker as a research impact statement. Presentations demonstrate visibility, but not necessarily software impact. It should be removed unless the authors connect those presentations to concrete research adoption or integration. Similarly, claims about possible future integrations should be removed, since they are not evidence of current demonstrated research impact.

Overall, I recommend major revisions. To better organize the paper, the authors can structure the paper around a clearer central argument: what Uno enables, establish that the capability is not already available in existing tools, how the software design supports that goal, and what existing evidence demonstrates its usefulness.
